@@ -1,10 +1,11 @@
-<?php namespace Granada;
+<?php
+
+namespace Granada;
 
 use ArrayAccess;
 use PDO;
 
 /**
- *
  * Granada
  *
  * Copyright (c) 2013, Erik Wiesenthal
@@ -48,61 +49,59 @@ use PDO;
  * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
  */
-
-class ORM implements ArrayAccess {
-
+class ORM implements ArrayAccess
+{
     // ----------------------- //
     // --- CLASS CONSTANTS --- //
     // ----------------------- //
 
     // WHERE and HAVING condition array keys
-    const CONDITION_FRAGMENT = 0;
-    const CONDITION_VALUES = 1;
+    public const CONDITION_FRAGMENT = 0;
+    public const CONDITION_VALUES   = 1;
 
-    const DEFAULT_CONNECTION = 'default';
+    public const DEFAULT_CONNECTION = 'default';
 
     // Limit clause style
-    const LIMIT_STYLE_TOP_N = "top";
-    const LIMIT_STYLE_LIMIT = "limit";
+    public const LIMIT_STYLE_TOP_N = 'top';
+    public const LIMIT_STYLE_LIMIT = 'limit';
 
     // ------------------------ //
     // --- CLASS PROPERTIES --- //
     // ------------------------ //
 
     // Class configuration
-    protected static $_default_config = array(
-        'connection_string' => 'sqlite::memory:',
-        'id_column' => 'id',
-        'id_column_overrides' => array(),
-        'error_mode' => PDO::ERRMODE_EXCEPTION,
-        'username' => null,
-        'password' => null,
-        'driver_options' => null,
-        'identifier_quote_character' => null, // if this is null, will be autodetected
-        'limit_clause_style' => null, // if this is null, will be autodetected
-        'logging' => false,
-        'logger' => null,
-        'caching' => false,
-        'return_result_sets' => true,
+    protected static $_default_config = [
+        'connection_string'           => 'sqlite::memory:',
+        'id_column'                   => 'id',
+        'id_column_overrides'         => [],
+        'error_mode'                  => PDO::ERRMODE_EXCEPTION,
+        'username'                    => null,
+        'password'                    => null,
+        'driver_options'              => null,
+        'identifier_quote_character'  => null, // if this is null, will be autodetected
+        'limit_clause_style'          => null, // if this is null, will be autodetected
+        'logging'                     => false,
+        'logger'                      => null,
+        'caching'                     => false,
+        'return_result_sets'          => true,
         'find_many_primary_id_as_key' => true,
-    );
+    ];
 
     // Map of configuration settings
-    protected static $_config = array();
+    protected static $_config = [];
 
     // Map of database connections, instances of the PDO class
-    protected static $_db = array();
+    protected static $_db = [];
 
     // Last query run, only populated if logging is enabled
     protected static $_last_query;
 
     // Log of all queries run, mapped by connection key, only populated if logging is enabled
-    protected static $_query_log = array();
+    protected static $_query_log = [];
 
     // Query cache, only used if query caching is enabled
-    protected static $_query_cache = array();
+    protected static $_query_cache = [];
 
     // Reference to previously used PDOStatement object to enable low-level access, if needed
     protected static $_last_statement = null;
@@ -121,16 +120,16 @@ class ORM implements ArrayAccess {
     protected $_table_alias = null;
 
     // Values to be bound to the query
-    protected $_values = array();
+    protected $_values = [];
 
     // Columns to select in the result
-    protected $_result_columns = array('*');
+    protected $_result_columns = ['*'];
 
     // Are we using the default result column or have these been manually changed?
     protected $_using_default_result_columns = true;
 
     // Join sources
-    protected $_join_sources = array();
+    protected $_join_sources = [];
 
     // Should the query include a DISTINCT keyword?
     protected $_distinct = false;
@@ -142,11 +141,11 @@ class ORM implements ArrayAccess {
     protected $_raw_query = '';
 
     // The raw query parameters
-    protected $_raw_parameters = array();
+    protected $_raw_parameters = [];
 
     // Array of WHERE clauses
-    protected $_where_conditions = array();
-    protected $_where_conditions_stash = array();
+    protected $_where_conditions       = [];
+    protected $_where_conditions_stash = [];
 
     // LIMIT
     protected $_limit = null;
@@ -155,26 +154,26 @@ class ORM implements ArrayAccess {
     protected $_offset = null;
 
     // ORDER BY
-    protected $_order_by = array();
+    protected $_order_by = [];
 
     // GROUP BY
-    protected $_group_by = array();
+    protected $_group_by = [];
 
     // HAVING
-    protected $_having_conditions = array();
+    protected $_having_conditions = [];
 
     // The data for a hydrated instance of the class
-    protected $_data = array();
+    protected $_data = [];
 
     // Fields that have been modified during the
     // lifetime of the object
-    protected $_dirty_fields = array();
+    protected $_dirty_fields = [];
 
     // The data as at hydration time, used for comparison of dirty fields
-    protected $_clean_data = array();
+    protected $_clean_data = [];
 
     // Fields that are to be inserted in the DB raw
-    protected $_expr_fields = array();
+    protected $_expr_fields = [];
 
     // Is this a new object (has create() been called)?
     protected $_is_new = false;
@@ -188,7 +187,6 @@ class ORM implements ArrayAccess {
 
     // associative results flag
     protected $_associative_results = true;
-
 
     // ---------------------- //
     // --- STATIC METHODS --- //
@@ -207,8 +205,9 @@ class ORM implements ArrayAccess {
      * @param mixed $value
      * @param string $connection_name Which connection to use
      */
-    public static function configure($key, $value = null, $connection_name = self::DEFAULT_CONNECTION) {
-        self::_setup_db_config($connection_name); //ensures at least default config is set
+    public static function configure($key, $value = null, $connection_name = self::DEFAULT_CONNECTION)
+    {
+        self::_setup_db_config($connection_name); // ensures at least default config is set
 
         if (is_array($key)) {
             // Shortcut: If only one array argument is passed,
@@ -221,7 +220,7 @@ class ORM implements ArrayAccess {
                 // Shortcut: If only one string argument is passed,
                 // assume it's a connection string
                 $value = $key;
-                $key = 'connection_string';
+                $key   = 'connection_string';
             }
             self::$_config[$connection_name][$key] = $value;
         }
@@ -232,19 +231,21 @@ class ORM implements ArrayAccess {
      * @param string $key
      * @param string $connection_name Which connection to use
      */
-    public static function get_config($key = null, $connection_name = self::DEFAULT_CONNECTION) {
+    public static function get_config($key = null, $connection_name = self::DEFAULT_CONNECTION)
+    {
         if ($key) {
             return self::$_config[$connection_name][$key];
-        } else {
-            return self::$_config[$connection_name];
         }
+
+        return self::$_config[$connection_name];
     }
 
     /**
      * Delete all configs in _config array.
      */
-    public static function reset_config() {
-        self::$_config = array();
+    public static function reset_config()
+    {
+        self::$_config = [];
     }
 
     /**
@@ -257,18 +258,23 @@ class ORM implements ArrayAccess {
      * @param string $connection_name Which connection to use
      * @return ORM
      */
-    public static function for_table($table_name, $connection_name = self::DEFAULT_CONNECTION) {
+    public static function for_table($table_name, $connection_name = self::DEFAULT_CONNECTION)
+    {
         self::_setup_db($connection_name);
-        return new self($table_name, array(), $connection_name);
+
+        return new self($table_name, [], $connection_name);
     }
 
     /**
      * Set up the database connection used by the class
      * @param string $connection_name Which connection to use
      */
-    protected static function _setup_db($connection_name = self::DEFAULT_CONNECTION) {
-        if (!array_key_exists($connection_name, self::$_db) ||
-            !is_object(self::$_db[$connection_name])) {
+    protected static function _setup_db($connection_name = self::DEFAULT_CONNECTION)
+    {
+        if (
+            !array_key_exists($connection_name, self::$_db) ||
+            !is_object(self::$_db[$connection_name])
+        ) {
             self::_setup_db_config($connection_name);
 
             $db = new PDO(
@@ -283,11 +289,12 @@ class ORM implements ArrayAccess {
         }
     }
 
-   /**
-    * Ensures configuration (mulitple connections) is at least set to default.
-    * @param string $connection_name Which connection to use
-    */
-    protected static function _setup_db_config($connection_name) {
+    /**
+     * Ensures configuration (mulitple connections) is at least set to default.
+     * @param string $connection_name Which connection to use
+     */
+    protected static function _setup_db_config($connection_name)
+    {
         if (!array_key_exists($connection_name, self::$_config)) {
             self::$_config[$connection_name] = self::$_default_config;
         }
@@ -301,7 +308,8 @@ class ORM implements ArrayAccess {
      * @param PDO $db
      * @param string $connection_name Which connection to use
      */
-    public static function set_db($db, $connection_name = self::DEFAULT_CONNECTION) {
+    public static function set_db($db, $connection_name = self::DEFAULT_CONNECTION)
+    {
         self::_setup_db_config($connection_name);
         self::$_db[$connection_name] = $db;
         self::_setup_identifier_quote_character($connection_name);
@@ -311,9 +319,10 @@ class ORM implements ArrayAccess {
     /**
      * Close and delete all registered PDO objects in _db array.
      */
-    public static function reset_db() {
+    public static function reset_db()
+    {
         self::$_db = null;
-        self::$_db = array();
+        self::$_db = [];
     }
 
     /**
@@ -323,10 +332,10 @@ class ORM implements ArrayAccess {
      * this will do nothing.
      * @param string $connection_name Which connection to use
      */
-    protected static function _setup_identifier_quote_character($connection_name) {
+    protected static function _setup_identifier_quote_character($connection_name)
+    {
         if (is_null(self::$_config[$connection_name]['identifier_quote_character'])) {
-            self::$_config[$connection_name]['identifier_quote_character'] =
-                self::_detect_identifier_quote_character($connection_name);
+            self::$_config[$connection_name]['identifier_quote_character'] = self::_detect_identifier_quote_character($connection_name);
         }
     }
 
@@ -336,10 +345,10 @@ class ORM implements ArrayAccess {
      * ORM::configure('limit_clause_style', 'top'), this will do nothing.
      * @param string $connection_name Which connection to use
      */
-    public static function _setup_limit_clause_style($connection_name) {
+    public static function _setup_limit_clause_style($connection_name)
+    {
         if (is_null(self::$_config[$connection_name]['limit_clause_style'])) {
-            self::$_config[$connection_name]['limit_clause_style'] =
-                self::_detect_limit_clause_style($connection_name);
+            self::$_config[$connection_name]['limit_clause_style'] = self::_detect_limit_clause_style($connection_name);
         }
     }
 
@@ -349,8 +358,9 @@ class ORM implements ArrayAccess {
      * @param string $connection_name Which connection to use
      * @return string
      */
-    protected static function _detect_identifier_quote_character($connection_name) {
-        switch(self::get_db($connection_name)->getAttribute(PDO::ATTR_DRIVER_NAME)) {
+    protected static function _detect_identifier_quote_character($connection_name)
+    {
+        switch (self::get_db($connection_name)->getAttribute(PDO::ATTR_DRIVER_NAME)) {
             case 'pgsql':
             case 'sqlsrv':
             case 'dblib':
@@ -358,6 +368,7 @@ class ORM implements ArrayAccess {
             case 'sybase':
             case 'firebird':
                 return '"';
+
             case 'mysql':
             case 'sqlite':
             case 'sqlite2':
@@ -372,12 +383,14 @@ class ORM implements ArrayAccess {
      * @param string $connection_name Which connection to use
      * @return string Limit clause style keyword/constant
      */
-    protected static function _detect_limit_clause_style($connection_name) {
-        switch(self::get_db($connection_name)->getAttribute(PDO::ATTR_DRIVER_NAME)) {
+    protected static function _detect_limit_clause_style($connection_name)
+    {
+        switch (self::get_db($connection_name)->getAttribute(PDO::ATTR_DRIVER_NAME)) {
             case 'sqlsrv':
             case 'dblib':
             case 'mssql':
                 return ORM::LIMIT_STYLE_TOP_N;
+
             default:
                 return ORM::LIMIT_STYLE_LIMIT;
         }
@@ -391,8 +404,10 @@ class ORM implements ArrayAccess {
      * @param string $connection_name Which connection to use
      * @return PDO
      */
-    public static function get_db($connection_name = self::DEFAULT_CONNECTION) {
+    public static function get_db($connection_name = self::DEFAULT_CONNECTION)
+    {
         self::_setup_db($connection_name); // required in case this is called before Idiorm is instantiated
+
         return self::$_db[$connection_name];
     }
 
@@ -407,8 +422,10 @@ class ORM implements ArrayAccess {
      * @param string $connection_name Which connection to use
      * @return bool Success
      */
-    public static function raw_execute($query, $parameters = array(), $connection_name = self::DEFAULT_CONNECTION) {
+    public static function raw_execute($query, $parameters = [], $connection_name = self::DEFAULT_CONNECTION)
+    {
         self::_setup_db($connection_name);
+
         return self::_execute($query, $parameters, $connection_name);
     }
 
@@ -417,20 +434,22 @@ class ORM implements ArrayAccess {
      * Useful for access to PDOStatement::rowCount() or error information
      * @return \PDOStatement
      */
-    public static function get_last_statement() {
+    public static function get_last_statement()
+    {
         return self::$_last_statement;
     }
 
-   /**
-    * Internal helper method for executing statments. Logs queries, and
-    * stores statement object in ::_last_statment, accessible publicly
-    * through ::get_last_statement()
-    * @param string $query
-    * @param array $parameters An array of parameters to be bound in to the query
-    * @param string $connection_name Which connection to use
-    * @return bool Response of PDOStatement::execute()
-    */
-    protected static function _execute($query, $parameters = array(), $connection_name = self::DEFAULT_CONNECTION) {
+    /**
+     * Internal helper method for executing statments. Logs queries, and
+     * stores statement object in ::_last_statment, accessible publicly
+     * through ::get_last_statement()
+     * @param string $query
+     * @param array $parameters An array of parameters to be bound in to the query
+     * @param string $connection_name Which connection to use
+     * @return bool Response of PDOStatement::execute()
+     */
+    protected static function _execute($query, $parameters = [], $connection_name = self::DEFAULT_CONNECTION)
+    {
         self::_log_query($query, $parameters, $connection_name);
         $statement = self::get_db($connection_name)->prepare($query);
 
@@ -452,31 +471,31 @@ class ORM implements ArrayAccess {
      * @param string $connection_name Which connection to use
      * @return bool
      */
-    protected static function _log_query($query, $parameters, $connection_name) {
+    protected static function _log_query($query, $parameters, $connection_name)
+    {
         // If logging is not enabled, do nothing
         if (!self::$_config[$connection_name]['logging']) {
             return false;
         }
 
         if (!isset(self::$_query_log[$connection_name])) {
-            self::$_query_log[$connection_name] = array();
+            self::$_query_log[$connection_name] = [];
         }
 
         if (count($parameters) > 0) {
             // Escape the parameters it not null
-            foreach ($parameters as $key => $parameter)
-            {
+            foreach ($parameters as $key => $parameter) {
                 $parameters[$key] = ($parameter === null) ? 'NULL' : self::$_db[$connection_name]->quote($parameter);
             }
 
             // Avoid %format collision for vsprintf
-            $query = str_replace("%", "%%", $query);
+            $query = str_replace('%', '%%', $query);
 
             // Replace placeholders in the query for vsprintf
-            if(false !== strpos($query, "'") || false !== strpos($query, '"')) {
-                $query = Orm\Str::str_replace_outside_quotes("?", "%s", $query);
+            if (false !== strpos($query, "'") || false !== strpos($query, '"')) {
+                $query = Orm\Str::str_replace_outside_quotes('?', '%s', $query);
             } else {
-                $query = str_replace("?", "%s", $query);
+                $query = str_replace('?', '%s', $query);
             }
 
             // Replace the question marks in the query with the parameters
@@ -485,11 +504,10 @@ class ORM implements ArrayAccess {
             $bound_query = $query;
         }
 
-        self::$_last_query = $bound_query;
+        self::$_last_query                    = $bound_query;
         self::$_query_log[$connection_name][] = $bound_query;
 
-
-        if(is_callable(self::$_config[$connection_name]['logger'])){
+        if (is_callable(self::$_config[$connection_name]['logger'])) {
             $logger = self::$_config[$connection_name]['logger'];
             $logger($bound_query);
         }
@@ -505,7 +523,8 @@ class ORM implements ArrayAccess {
      * @param null|string $connection_name Which connection to use
      * @return string
      */
-    public static function get_last_query($connection_name = null) {
+    public static function get_last_query($connection_name = null)
+    {
         if ($connection_name === null) {
             return self::$_last_query;
         }
@@ -523,18 +542,21 @@ class ORM implements ArrayAccess {
      * set to true. Otherwise, returned array will be empty.
      * @param string $connection_name Which connection to use
      */
-    public static function get_query_log($connection_name = self::DEFAULT_CONNECTION) {
+    public static function get_query_log($connection_name = self::DEFAULT_CONNECTION)
+    {
         if (isset(self::$_query_log[$connection_name])) {
             return self::$_query_log[$connection_name];
         }
-        return array();
+
+        return [];
     }
 
     /**
      * Get a list of the available connection names
      * @return array
      */
-    public static function get_connection_names() {
+    public static function get_connection_names()
+    {
         return array_keys(self::$_db);
     }
 
@@ -546,15 +568,16 @@ class ORM implements ArrayAccess {
      * "Private" constructor; shouldn't be called directly.
      * Use the ORM::for_table factory method instead.
      */
-    protected function __construct($table_name, $data = array(), $connection_name = self::DEFAULT_CONNECTION) {
+    protected function __construct($table_name, $data = [], $connection_name = self::DEFAULT_CONNECTION)
+    {
         $this->_table_name = $table_name;
-        $this->_data = $data;
-        $this->_clean_data = array();
+        $this->_data       = $data;
+        $this->_clean_data = [];
 
         $this->_connection_name = $connection_name;
 
         // Set the flag as config dictates
-        $this->_associative_results  = self::$_config[$this->_connection_name]['find_many_primary_id_as_key'];
+        $this->_associative_results = self::$_config[$this->_connection_name]['find_many_primary_id_as_key'];
 
         self::_setup_db_config($connection_name);
     }
@@ -567,11 +590,13 @@ class ORM implements ArrayAccess {
      * dirty so all will be saved to the database when
      * save() is called.
      */
-    public function create($data=null) {
+    public function create($data = null)
+    {
         $this->_is_new = true;
         if (!is_null($data)) {
             return $this->hydrate($data)->force_all_dirty();
         }
+
         return $this;
     }
 
@@ -579,8 +604,10 @@ class ORM implements ArrayAccess {
      * Set the ORM instance to return non associative results sets
      * @return ORM instance
      */
-    public function non_associative() {
+    public function non_associative()
+    {
         $this->_associative_results = false;
+
         return $this;
     }
 
@@ -588,8 +615,10 @@ class ORM implements ArrayAccess {
      * Set the ORM instance to return associative results sets
      * @return ORM instance
      */
-    public function associative() {
+    public function associative()
+    {
         $this->_associative_results = true;
+
         return $this;
     }
 
@@ -597,8 +626,10 @@ class ORM implements ArrayAccess {
      * Set the ORM instance to return associative (or not) results sets, as config dictates
      * @return ORM instance
      */
-    public function reset_associative() {
-        $this->_associative_results  = self::$_config[$this->_connection_name]['find_many_primary_id_as_key'];
+    public function reset_associative()
+    {
+        $this->_associative_results = self::$_config[$this->_connection_name]['find_many_primary_id_as_key'];
+
         return $this;
     }
 
@@ -610,8 +641,10 @@ class ORM implements ArrayAccess {
      * not normally be used in manually built queries. If you don't know why
      * you would want to use this, you should probably just ignore it.
      */
-    public function use_id_column($id_column) {
+    public function use_id_column($id_column)
+    {
         $this->_instance_id_column = $id_column;
+
         return $this;
     }
 
@@ -619,10 +652,12 @@ class ORM implements ArrayAccess {
      * Create an ORM instance from the given row (an associative
      * array of data fetched from the database)
      */
-    protected function _create_instance_from_row($row) {
+    protected function _create_instance_from_row($row)
+    {
         $instance = static::for_table($this->_table_name, $this->_connection_name);
         $instance->use_id_column($this->_instance_id_column);
         $instance->hydrate($row);
+
         return $instance;
     }
 
@@ -635,7 +670,8 @@ class ORM implements ArrayAccess {
      * to this method. This will perform a primary key
      * lookup on the table.
      */
-    public function find_one($id=null) {
+    public function find_one($id = null)
+    {
         if (!is_null($id)) {
             $this->where_id_is($id);
         }
@@ -656,10 +692,12 @@ class ORM implements ArrayAccess {
      * no rows were returned.
      * @return array|ResultSet
      */
-    public function find_many() {
-        if(self::$_config[$this->_connection_name]['return_result_sets']) {
+    public function find_many()
+    {
+        if (self::$_config[$this->_connection_name]['return_result_sets']) {
             return $this->find_result_set();
         }
+
         return $this->_find_many();
     }
 
@@ -667,7 +705,8 @@ class ORM implements ArrayAccess {
      * Perform a find_many then map the results through a function
      * @param callable $func
      */
-    public function find_map($func) {
+    public function find_map($func)
+    {
         return array_map($func, $this->find_many()->as_array());
     }
 
@@ -676,14 +715,16 @@ class ORM implements ArrayAccess {
      * @param string $connection_name
      * @return string
      */
-    public function get_select_query($connection_name = self::DEFAULT_CONNECTION) {
+    public function get_select_query($connection_name = self::DEFAULT_CONNECTION)
+    {
         // Ensure logging works
-        $before_log = self::$_config[$connection_name]['logging'];
+        $before_log                                 = self::$_config[$connection_name]['logging'];
         self::$_config[$connection_name]['logging'] = true;
 
         $this->_log_query($this->_build_select(), $this->_values, $connection_name);
 
         self::$_config[$connection_name]['logging'] = $before_log;
+
         return $this->get_last_query();
     }
 
@@ -694,8 +735,10 @@ class ORM implements ArrayAccess {
      * no rows were returned.
      * @return array
      */
-    protected function _find_many($associative = true) {
+    protected function _find_many($associative = true)
+    {
         $rows = $this->_run();
+
         return $this->_get_instances($rows);
     }
 
@@ -706,14 +749,16 @@ class ORM implements ArrayAccess {
      * @param array $rows
      * @return array
      */
-    protected function _get_instances($rows) {
-        $size = count($rows);
-        $instances = array();
+    protected function _get_instances($rows)
+    {
+        $size      = count($rows);
+        $instances = [];
         for ($i = 0; $i < $size; $i++) {
-            $row = $this->_create_instance_from_row($rows[$i]);
-            $key = (isset($row->{$this->_instance_id_column}) && $this->_associative_results && $row->id()) ? $row->id() : $i;
+            $row             = $this->_create_instance_from_row($rows[$i]);
+            $key             = (isset($row->{$this->_instance_id_column}) && $this->_associative_results && $row->id()) ? $row->id() : $i;
             $instances[$key] = $row;
         }
+
         return $instances;
     }
 
@@ -723,14 +768,15 @@ class ORM implements ArrayAccess {
      * containing instances of the ORM class.
      * @return \Granada\ResultSet
      */
-    public function find_result_set() {
+    public function find_result_set()
+    {
         $resultSetClass = $this->resultSetClass;
-        if(is_a($resultSetClass, 'Granada\ResultSet', true)){
+        if (is_a($resultSetClass, 'Granada\ResultSet', true)) {
             $result = new $resultSetClass($this->_find_many());
-        }
-        else{
+        } else {
             $result = new ResultSet($this->_find_many());
         }
+
         return $result;
     }
 
@@ -740,7 +786,8 @@ class ORM implements ArrayAccess {
      * or an empty array if no rows were returned.
      * @return array
      */
-    public function find_array() {
+    public function find_array()
+    {
         return $this->_run();
     }
 
@@ -749,7 +796,8 @@ class ORM implements ArrayAccess {
      * Will return an integer representing the number of
      * rows returned.
      */
-    public function count($column = '*') {
+    public function count($column = '*')
+    {
         return $this->_call_aggregate_db_function(__FUNCTION__, $column);
     }
 
@@ -758,7 +806,8 @@ class ORM implements ArrayAccess {
      * Will return the max value of the choosen column.
      * @param string $column
      */
-    public function max($column)  {
+    public function max($column)
+    {
         return $this->_call_aggregate_db_function(__FUNCTION__, $column);
     }
 
@@ -767,7 +816,8 @@ class ORM implements ArrayAccess {
      * Will return the min value of the choosen column.
      * @param string $column
      */
-    public function min($column)  {
+    public function min($column)
+    {
         return $this->_call_aggregate_db_function(__FUNCTION__, $column);
     }
 
@@ -776,7 +826,8 @@ class ORM implements ArrayAccess {
      * Will return the average value of the choosen column.
      * @param string $column
      */
-    public function avg($column)  {
+    public function avg($column)
+    {
         return $this->_call_aggregate_db_function(__FUNCTION__, $column);
     }
 
@@ -785,7 +836,8 @@ class ORM implements ArrayAccess {
      * Will return the sum of the choosen column.
      * @param string $column
      */
-    public function sum($column)  {
+    public function sum($column)
+    {
         return $this->_call_aggregate_db_function(__FUNCTION__, $column);
     }
 
@@ -795,41 +847,44 @@ class ORM implements ArrayAccess {
      * @param string $column The column to execute the aggregate query against
      * @return int
      */
-    protected function _call_aggregate_db_function($sql_function, $column) {
-        $alias = strtolower($sql_function);
+    protected function _call_aggregate_db_function($sql_function, $column)
+    {
+        $alias        = strtolower($sql_function);
         $sql_function = strtoupper($sql_function);
-        if('*' != $column) {
+        if ('*' != $column) {
             $column = $this->_quote_identifier($column);
         }
-        $result_columns = $this->_result_columns;
-        $this->_result_columns = array();
+        $result_columns        = $this->_result_columns;
+        $this->_result_columns = [];
         $this->select_expr("$sql_function($column)", $alias);
-        $result = $this->find_one();
+        $result                = $this->find_one();
         $this->_result_columns = $result_columns;
 
         $return_value = 0;
-        if($result !== false && isset($result->$alias)) {
+        if ($result !== false && isset($result->$alias)) {
             if (!is_numeric($result->$alias)) {
                 $return_value = $result->$alias;
-            }
-            elseif((int) $result->$alias == (float) $result->$alias) {
+            } elseif ((int) $result->$alias == (float) $result->$alias) {
                 $return_value = (int) $result->$alias;
             } else {
-                 $return_value = (float) $result->$alias;
+                $return_value = (float) $result->$alias;
             }
         }
+
         return $return_value;
     }
 
-     /**
+    /**
      * This method can be called to hydrate (populate) this
      * instance of the class from an associative array of data.
      * This will usually be called only from inside the class,
      * but it's public in case you need to call it directly.
      */
-    public function hydrate($data=array()) {
-        $this->_data = $data;
-        $this->_clean_data = array();
+    public function hydrate($data = [])
+    {
+        $this->_data       = $data;
+        $this->_clean_data = [];
+
         return $this;
     }
 
@@ -837,8 +892,10 @@ class ORM implements ArrayAccess {
      * Force the ORM to flag all the fields in the $data array
      * as "dirty" and therefore update them when save() is called.
      */
-    public function force_all_dirty() {
+    public function force_all_dirty()
+    {
         $this->_dirty_fields = $this->_data;
+
         return $this;
     }
 
@@ -850,10 +907,12 @@ class ORM implements ArrayAccess {
      * is called, all other query building methods will be ignored.
      * @param string $query
      */
-    public function raw_query($query, $parameters = array()) {
-        $this->_is_raw_query = true;
-        $this->_raw_query = $query;
+    public function raw_query($query, $parameters = [])
+    {
+        $this->_is_raw_query   = true;
+        $this->_raw_query      = $query;
         $this->_raw_parameters = $parameters;
+
         return $this;
     }
 
@@ -861,8 +920,10 @@ class ORM implements ArrayAccess {
      * Add an alias for the main table to be used in SELECT queries
      * @param string $alias
      */
-    public function table_alias($alias) {
+    public function table_alias($alias)
+    {
         $this->_table_alias = $alias;
+
         return $this;
     }
 
@@ -871,17 +932,19 @@ class ORM implements ArrayAccess {
      * of columns returned by the SELECT query. The second optional
      * argument is the alias to return the expression as.
      */
-    protected function _add_result_column($expr, $alias=null) {
+    protected function _add_result_column($expr, $alias = null)
+    {
         if (!is_null($alias)) {
-            $expr .= " AS " . $this->_quote_identifier($alias);
+            $expr .= ' AS ' . $this->_quote_identifier($alias);
         }
 
         if ($this->_using_default_result_columns) {
-            $this->_result_columns = array($expr);
+            $this->_result_columns               = [$expr];
             $this->_using_default_result_columns = false;
         } else {
             $this->_result_columns[] = $expr;
         }
+
         return $this;
     }
 
@@ -891,12 +954,14 @@ class ORM implements ArrayAccess {
      * the alias to return the column as.
      * @param string $alias
      */
-    public function select($column, $alias=null) {
-        $columns = array_map('trim',explode(',',$column));
-        foreach($columns as $column){
+    public function select($column, $alias = null)
+    {
+        $columns = array_map('trim', explode(',', $column));
+        foreach ($columns as $column) {
             $column = $this->_quote_identifier($column);
             $this->_add_result_column($column, $alias);
         }
+
         return $this;
     }
 
@@ -905,7 +970,8 @@ class ORM implements ArrayAccess {
      * by the SELECT query. The second optional argument is
      * the alias to return the column as.
      */
-    public function select_expr($expr, $alias=null) {
+    public function select_expr($expr, $alias = null)
+    {
         return $this->_add_result_column($expr, $alias);
     }
 
@@ -923,17 +989,19 @@ class ORM implements ArrayAccess {
      *
      * @return ORM
      */
-    public function select_many() {
+    public function select_many()
+    {
         $columns = func_get_args();
-        if(!empty($columns)) {
+        if (!empty($columns)) {
             $columns = $this->_normalise_select_many_columns($columns);
-            foreach($columns as $alias => $column) {
-                if(is_numeric($alias)) {
+            foreach ($columns as $alias => $column) {
+                if (is_numeric($alias)) {
                     $alias = null;
                 }
                 $this->select($column, $alias);
             }
         }
+
         return $this;
     }
 
@@ -951,17 +1019,19 @@ class ORM implements ArrayAccess {
      *
      * @return ORM
      */
-    public function select_many_expr() {
+    public function select_many_expr()
+    {
         $columns = func_get_args();
-        if(!empty($columns)) {
+        if (!empty($columns)) {
             $columns = $this->_normalise_select_many_columns($columns);
-            foreach($columns as $alias => $column) {
-                if(is_numeric($alias)) {
+            foreach ($columns as $alias => $column) {
+                if (is_numeric($alias)) {
                     $alias = null;
                 }
                 $this->select_expr($column, $alias);
             }
         }
+
         return $this;
     }
 
@@ -976,12 +1046,13 @@ class ORM implements ArrayAccess {
      * @param array $columns
      * @return array
      */
-    protected function _normalise_select_many_columns($columns) {
-        $return = array();
-        foreach($columns as $column) {
-            if(is_array($column)) {
-                foreach($column as $key => $value) {
-                    if(!is_numeric($key)) {
+    protected function _normalise_select_many_columns($columns)
+    {
+        $return = [];
+        foreach ($columns as $column) {
+            if (is_array($column)) {
+                foreach ($column as $key => $value) {
+                    if (!is_numeric($key)) {
                         $return[$key] = $value;
                     } else {
                         $return[] = $value;
@@ -991,14 +1062,17 @@ class ORM implements ArrayAccess {
                 $return[] = $column;
             }
         }
+
         return $return;
     }
 
     /**
      * Add a DISTINCT keyword before the list of columns in the SELECT query
      */
-    public function distinct() {
+    public function distinct()
+    {
         $this->_distinct = true;
+
         return $this;
     }
 
@@ -1025,8 +1099,8 @@ class ORM implements ArrayAccess {
      * The final (optional) argument specifies an alias for the joined table.
      * @param string $join_operator
      */
-    protected function _add_join_source($join_operator, $table, $constraint, $table_alias=null) {
-
+    protected function _add_join_source($join_operator, $table, $constraint, $table_alias = null)
+    {
         $join_operator = trim("{$join_operator} JOIN");
 
         $table = $this->_quote_identifier($table);
@@ -1039,13 +1113,14 @@ class ORM implements ArrayAccess {
 
         // Build the constraint
         if (is_array($constraint)) {
-            list($first_column, $operator, $second_column) = $constraint;
-            $first_column = $this->_quote_identifier($first_column);
-            $second_column = $this->_quote_identifier($second_column);
-            $constraint = "{$first_column} {$operator} {$second_column}";
+            [$first_column, $operator, $second_column] = $constraint;
+            $first_column                              = $this->_quote_identifier($first_column);
+            $second_column                             = $this->_quote_identifier($second_column);
+            $constraint                                = "{$first_column} {$operator} {$second_column}";
         }
 
         $this->_join_sources[] = "{$join_operator} {$table} ON {$constraint}";
+
         return $this;
     }
 
@@ -1053,8 +1128,9 @@ class ORM implements ArrayAccess {
      * Add a simple JOIN source to the query
      * @param string $table_alias
      */
-    public function join($table, $constraint, $table_alias=null) {
-        return $this->_add_join_source("", $table, $constraint, $table_alias);
+    public function join($table, $constraint, $table_alias = null)
+    {
+        return $this->_add_join_source('', $table, $constraint, $table_alias);
     }
 
     /**
@@ -1062,8 +1138,9 @@ class ORM implements ArrayAccess {
      * @param string $table
      * @param string[] $constraint
      */
-    public function inner_join($table, $constraint, $table_alias=null) {
-        return $this->_add_join_source("INNER", $table, $constraint, $table_alias);
+    public function inner_join($table, $constraint, $table_alias = null)
+    {
+        return $this->_add_join_source('INNER', $table, $constraint, $table_alias);
     }
 
     /**
@@ -1071,8 +1148,9 @@ class ORM implements ArrayAccess {
      * @param string $table
      * @param string[] $constraint
      */
-    public function left_outer_join($table, $constraint, $table_alias=null) {
-        return $this->_add_join_source("LEFT OUTER", $table, $constraint, $table_alias);
+    public function left_outer_join($table, $constraint, $table_alias = null)
+    {
+        return $this->_add_join_source('LEFT OUTER', $table, $constraint, $table_alias);
     }
 
     /**
@@ -1080,8 +1158,9 @@ class ORM implements ArrayAccess {
      * @param string $table
      * @param string[] $constraint
      */
-    public function right_outer_join($table, $constraint, $table_alias=null) {
-        return $this->_add_join_source("RIGHT OUTER", $table, $constraint, $table_alias);
+    public function right_outer_join($table, $constraint, $table_alias = null)
+    {
+        return $this->_add_join_source('RIGHT OUTER', $table, $constraint, $table_alias);
     }
 
     /**
@@ -1089,14 +1168,16 @@ class ORM implements ArrayAccess {
      * @param string $table
      * @param string[] $constraint
      */
-    public function full_outer_join($table, $constraint, $table_alias=null) {
-        return $this->_add_join_source("FULL OUTER", $table, $constraint, $table_alias);
+    public function full_outer_join($table, $constraint, $table_alias = null)
+    {
+        return $this->_add_join_source('FULL OUTER', $table, $constraint, $table_alias);
     }
 
     /**
      * Internal method to add a HAVING condition to the query
      */
-    protected function _add_having($fragment, $values=array()) {
+    protected function _add_having($fragment, $values = [])
+    {
         return $this->_add_condition('having', $fragment, $values);
     }
 
@@ -1104,14 +1185,16 @@ class ORM implements ArrayAccess {
      * Internal method to add a HAVING condition to the query
      * @param string $separator
      */
-    protected function _add_simple_having($column_name, $separator, $value) {
+    protected function _add_simple_having($column_name, $separator, $value)
+    {
         return $this->_add_simple_condition('having', $column_name, $separator, $value);
     }
 
     /**
      * Internal method to add a WHERE condition to the query
      */
-    protected function _add_where($fragment, $values=array()) {
+    protected function _add_where($fragment, $values = [])
+    {
         return $this->_add_condition('where', $fragment, $values);
     }
 
@@ -1119,22 +1202,25 @@ class ORM implements ArrayAccess {
      * Internal method to add a WHERE condition to the query
      * @param string $separator
      */
-    protected function _add_simple_where($column_name, $separator, $value) {
+    protected function _add_simple_where($column_name, $separator, $value)
+    {
         return $this->_add_simple_condition('where', $column_name, $separator, $value);
     }
 
     /**
      * Internal method to add a HAVING or WHERE condition to the query
      */
-    protected function _add_condition($type, $fragment, $values=array()) {
+    protected function _add_condition($type, $fragment, $values = [])
+    {
         $conditions_class_property_name = "_{$type}_conditions";
         if (!is_array($values)) {
-            $values = array($values);
+            $values = [$values];
         }
-        array_push($this->$conditions_class_property_name, array(
+        array_push($this->$conditions_class_property_name, [
             self::CONDITION_FRAGMENT => $fragment,
-            self::CONDITION_VALUES => $values,
-        ));
+            self::CONDITION_VALUES   => $values,
+        ]);
+
         return $this;
     }
 
@@ -1144,8 +1230,10 @@ class ORM implements ArrayAccess {
      *
      * @return ORM
      */
-    public function stash_where() {
+    public function stash_where()
+    {
         $this->_where_conditions_stash = $this->_where_conditions;
+
         return $this->clear_where();
     }
 
@@ -1154,40 +1242,47 @@ class ORM implements ArrayAccess {
      *
      * @return ORM
      */
-    public function pop_where() {
+    public function pop_where()
+    {
         foreach ($this->_where_conditions_stash as $stash) {
             $this->_where_conditions[] = $stash;
         }
-        $this->_where_conditions_stash = array();
+        $this->_where_conditions_stash = [];
+
         return $this;
     }
 
     /**
      * Clear / Reset the WHERE clause(s)
-	 * @return ORM
+     * @return ORM
      */
-    public function clear_where() {
-        $this->_where_conditions = array();
-		return $this;
+    public function clear_where()
+    {
+        $this->_where_conditions = [];
+
+        return $this;
     }
 
     /**
      * Clear / Reset the HAVING clause(s)
-	 * @return ORM
+     * @return ORM
      */
-    public function clear_having() {
-        $this->_having_conditions = array();
-		return $this;
+    public function clear_having()
+    {
+        $this->_having_conditions = [];
+
+        return $this;
     }
 
-   /**
+    /**
      * Helper method to compile a simple COLUMN SEPARATOR VALUE
      * style HAVING or WHERE condition into a string and value ready to
      * be passed to the _add_condition method. Avoids duplication
      * of the call to _quote_identifier
      * @param string $type
      */
-    protected function _add_simple_condition($type, $column_name, $separator, $value) {
+    protected function _add_simple_condition($type, $column_name, $separator, $value)
+    {
         // Add the table name in case of ambiguous columns
         if (count($this->_join_sources) > 0 && strpos($column_name, '.') === false) {
             $table = $this->_table_name;
@@ -1198,6 +1293,7 @@ class ORM implements ArrayAccess {
             $column_name = "{$table}.{$column_name}";
         }
         $column_name = $this->_quote_identifier($column_name);
+
         return $this->_add_condition($type, "{$column_name} {$separator} ?", $value);
     }
 
@@ -1205,17 +1301,19 @@ class ORM implements ArrayAccess {
      * Return a string containing the given number of question marks,
      * separated by commas. Eg "?, ?, ?"
      */
-    protected function _create_placeholders($fields) {
-        if(!empty($fields)) {
-            $db_fields = array();
-            foreach($fields as $key => $value) {
+    protected function _create_placeholders($fields)
+    {
+        if (!empty($fields)) {
+            $db_fields = [];
+            foreach ($fields as $key => $value) {
                 // Process expression fields directly into the query
-                if(array_key_exists($key, $this->_expr_fields)) {
+                if (array_key_exists($key, $this->_expr_fields)) {
                     $db_fields[] = $value;
                 } else {
                     $db_fields[] = '?';
                 }
             }
+
             return implode(', ', $db_fields);
         }
     }
@@ -1234,23 +1332,27 @@ class ORM implements ArrayAccess {
      * @param callable $callback
      * @return self
      */
-    public function onlyif($condition, $callback) {
+    public function onlyif($condition, $callback)
+    {
         if ($condition) {
             return $callback($this);
         }
 
         return $this;
     }
+
     /**
      * Add a WHERE column = value clause to your query. Each time
      * this is called in the chain, an additional WHERE will be
      * added, and these will be ANDed together when the final query
      * is built.
      */
-    public function where($column_name, $value) {
-		if (is_null($value)) {
-			return $this->where_null($column_name);
-		}
+    public function where($column_name, $value)
+    {
+        if (is_null($value)) {
+            return $this->where_null($column_name);
+        }
+
         return $this->where_equal($column_name, $value);
     }
 
@@ -1258,10 +1360,12 @@ class ORM implements ArrayAccess {
      * More explicitly named version of for the where() method.
      * Can be used if preferred.
      */
-    public function where_equal($column_name, $value) {
-		if (is_null($value)) {
-			return $this->where_null($column_name);
-		}
+    public function where_equal($column_name, $value)
+    {
+        if (is_null($value)) {
+            return $this->where_null($column_name);
+        }
+
         return $this->_add_simple_where($column_name, '=', $value);
     }
 
@@ -1270,17 +1374,20 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string|null $value
      */
-    public function where_not_equal($column_name, $value) {
-		if (is_null($value)) {
-			return $this->where_not_null($column_name);
-		}
+    public function where_not_equal($column_name, $value)
+    {
+        if (is_null($value)) {
+            return $this->where_not_null($column_name);
+        }
+
         return $this->_add_simple_where($column_name, '!=', $value);
     }
 
     /**
      * Special method to query the table by its primary key
      */
-    public function where_id_is($id) {
+    public function where_id_is($id)
+    {
         return $this->where($this->_get_id_column_name(), $id);
     }
 
@@ -1294,48 +1401,50 @@ class ORM implements ArrayAccess {
      *
      * Each condition will be ORed together when added to the final query.
      */
-    public function where_any_is($values, $operator='=') {
-        $data = array();
-        $query = array("((");
+    public function where_any_is($values, $operator = '=')
+    {
+        $data  = [];
+        $query = ['(('];
         $first = true;
         foreach ($values as $value) {
             if ($first) {
                 $first = false;
             } else {
-                $query[] = ") OR (";
+                $query[] = ') OR (';
             }
             $firstsub = true;
-            foreach($value as $key => $item) {
-                $op = is_string($operator) ? $operator : (isset($operator[$key]) ? $operator[$key] : '=');
+            foreach ($value as $key => $item) {
+                $op = is_string($operator) ? $operator : ($operator[$key] ?? '=');
                 if ($firstsub) {
                     $firstsub = false;
                 } else {
-                    $query[] = "AND";
+                    $query[] = 'AND';
                 }
                 $query[] = $this->_quote_identifier($key);
                 if (is_array($item)) {
                     $placeholders = $this->_create_placeholders($item);
-                    $data = array_merge($data, $item);
+                    $data         = array_merge($data, $item);
                     if ($op == '=') {
                         $query[] = 'IN (' . $placeholders . ')';
-                    } else if ($op == '!=') {
+                    } elseif ($op == '!=') {
                         $query[] = 'NOT IN (' . $placeholders . ')';
                     } else {
                         throw new \InvalidArgumentException('You only pass an array for = and !=.');
                     }
                 } else {
-					if (is_null($item) && ($op == '=')) {
-						$query[] = 'IS NULL';
-					} else if (is_null($item) && ($op == '!=')) {
-						$query[] = 'IS NOT NULL';
-					} else {
-						$query[] = $op . " ?";
-						$data[] = $item;
-					}
-				}
+                    if (is_null($item) && ($op == '=')) {
+                        $query[] = 'IS NULL';
+                    } elseif (is_null($item) && ($op == '!=')) {
+                        $query[] = 'IS NOT NULL';
+                    } else {
+                        $query[] = $op . ' ?';
+                        $data[]  = $item;
+                    }
+                }
             }
         }
-        $query[] = "))";
+        $query[] = '))';
+
         return $this->where_raw(implode(' ', $query), $data);
     }
 
@@ -1344,7 +1453,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string $value
      */
-    public function where_like($column_name, $value) {
+    public function where_like($column_name, $value)
+    {
         return $this->_add_simple_where($column_name, 'LIKE', $value);
     }
 
@@ -1353,7 +1463,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string $value
      */
-    public function where_not_like($column_name, $value) {
+    public function where_not_like($column_name, $value)
+    {
         return $this->_add_simple_where($column_name, 'NOT LIKE', $value);
     }
 
@@ -1362,7 +1473,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function where_gt($column_name, $value) {
+    public function where_gt($column_name, $value)
+    {
         return $this->_add_simple_where($column_name, '>', $value);
     }
 
@@ -1371,7 +1483,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function where_lt($column_name, $value) {
+    public function where_lt($column_name, $value)
+    {
         return $this->_add_simple_where($column_name, '<', $value);
     }
 
@@ -1380,7 +1493,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function where_gte($column_name, $value) {
+    public function where_gte($column_name, $value)
+    {
         return $this->_add_simple_where($column_name, '>=', $value);
     }
 
@@ -1389,7 +1503,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function where_lte($column_name, $value) {
+    public function where_lte($column_name, $value)
+    {
         return $this->_add_simple_where($column_name, '<=', $value);
     }
 
@@ -1398,7 +1513,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function where_lt_or_null($column_name, $value) {
+    public function where_lt_or_null($column_name, $value)
+    {
         return $this->where_raw('( ' . $this->_quote_identifier($column_name) . ' < ? OR ' . $this->_quote_identifier($column_name) . ' IS NULL )', $value);
     }
 
@@ -1407,7 +1523,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function where_lte_or_null($column_name, $value) {
+    public function where_lte_or_null($column_name, $value)
+    {
         return $this->where_raw('( ' . $this->_quote_identifier($column_name) . ' <= ? OR ' . $this->_quote_identifier($column_name) . ' IS NULL )', $value);
     }
 
@@ -1416,7 +1533,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function where_gt_or_null($column_name, $value) {
+    public function where_gt_or_null($column_name, $value)
+    {
         return $this->where_raw('( ' . $this->_quote_identifier($column_name) . ' > ? OR ' . $this->_quote_identifier($column_name) . ' IS NULL )', $value);
     }
 
@@ -1425,19 +1543,22 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function where_gte_or_null($column_name, $value) {
+    public function where_gte_or_null($column_name, $value)
+    {
         return $this->where_raw('( ' . $this->_quote_identifier($column_name) . ' >= ? OR ' . $this->_quote_identifier($column_name) . ' IS NULL )', $value);
     }
 
     /**
      * Add a WHERE ... IN clause to your query
      */
-    public function where_in($column_name, $values) {
+    public function where_in($column_name, $values)
+    {
         if (!$values) {
-            return $this->_add_where("0");
+            return $this->_add_where('0');
         }
-        $column_name = $this->_quote_identifier($column_name);
+        $column_name  = $this->_quote_identifier($column_name);
         $placeholders = $this->_create_placeholders($values);
+
         return $this->_add_where("{$column_name} IN ({$placeholders})", $values);
     }
 
@@ -1446,12 +1567,15 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string[] $values
      */
-    public function where_not_in($column_name, $values) {
+    public function where_not_in($column_name, $values)
+    {
         if ($values) {
-            $column_name = $this->_quote_identifier($column_name);
+            $column_name  = $this->_quote_identifier($column_name);
             $placeholders = $this->_create_placeholders($values);
+
             return $this->_add_where("{$column_name} NOT IN ({$placeholders})", $values);
         }
+
         return $this;
     }
 
@@ -1459,8 +1583,10 @@ class ORM implements ArrayAccess {
      * Add a WHERE column IS NULL clause to your query
      * @param string $column_name
      */
-    public function where_null($column_name) {
+    public function where_null($column_name)
+    {
         $column_name = $this->_quote_identifier($column_name);
+
         return $this->_add_where("{$column_name} IS NULL");
     }
 
@@ -1468,8 +1594,10 @@ class ORM implements ArrayAccess {
      * Add a WHERE column IS NOT NULL clause to your query
      * @param string $column_name
      */
-    public function where_not_null($column_name) {
+    public function where_not_null($column_name)
+    {
         $column_name = $this->_quote_identifier($column_name);
+
         return $this->_add_where("{$column_name} IS NOT NULL");
     }
 
@@ -1479,7 +1607,8 @@ class ORM implements ArrayAccess {
      * to the parameters supplied in the second argument.
      * @param string $clause
      */
-    public function where_raw($clause, $parameters=array()) {
+    public function where_raw($clause, $parameters = [])
+    {
         return $this->_add_where($clause, $parameters);
     }
 
@@ -1487,8 +1616,10 @@ class ORM implements ArrayAccess {
      * Add a LIMIT to the query
      * @param integer $limit
      */
-    public function limit($limit) {
+    public function limit($limit)
+    {
         $this->_limit = $limit;
+
         return $this;
     }
 
@@ -1496,8 +1627,10 @@ class ORM implements ArrayAccess {
      * Add an OFFSET to the query
      * @param integer $offset
      */
-    public function offset($offset) {
+    public function offset($offset)
+    {
         $this->_offset = $offset;
+
         return $this;
     }
 
@@ -1505,17 +1638,21 @@ class ORM implements ArrayAccess {
      * Add an ORDER BY clause to the query
      * @param string $ordering
      */
-    protected function _add_order_by($column_name, $ordering) {
-        $column_name = $this->_quote_identifier($column_name);
+    protected function _add_order_by($column_name, $ordering)
+    {
+        $column_name       = $this->_quote_identifier($column_name);
         $this->_order_by[] = "{$column_name} {$ordering}";
+
         return $this;
     }
 
     /**
      * Clear all ORDER BY clauses, used to override previous orders
      */
-    public function order_by_clear() {
-        $this->_order_by = array();
+    public function order_by_clear()
+    {
+        $this->_order_by = [];
+
         return $this;
     }
 
@@ -1523,7 +1660,8 @@ class ORM implements ArrayAccess {
      * Add an ORDER BY column DESC clause
      * @param string $column_name
      */
-    public function order_by_desc($column_name) {
+    public function order_by_desc($column_name)
+    {
         return $this->_add_order_by($column_name, 'DESC');
     }
 
@@ -1531,7 +1669,8 @@ class ORM implements ArrayAccess {
      * Add an ORDER BY column ASC clause
      * @param boolean|string $column_name
      */
-    public function order_by_asc($column_name) {
+    public function order_by_asc($column_name)
+    {
         return $this->_add_order_by($column_name, 'ASC');
     }
 
@@ -1539,8 +1678,10 @@ class ORM implements ArrayAccess {
      * Add an unquoted expression as an ORDER BY clause
      * @param string $clause
      */
-    public function order_by_expr($clause) {
+    public function order_by_expr($clause)
+    {
         $this->_order_by[] = $clause;
+
         return $this;
     }
 
@@ -1548,9 +1689,11 @@ class ORM implements ArrayAccess {
      * Add a column to the list of columns to GROUP BY
      * @param string $column_name
      */
-    public function group_by($column_name) {
-        $column_name = $this->_quote_identifier($column_name);
+    public function group_by($column_name)
+    {
+        $column_name       = $this->_quote_identifier($column_name);
         $this->_group_by[] = $column_name;
+
         return $this;
     }
 
@@ -1558,8 +1701,10 @@ class ORM implements ArrayAccess {
      * Add an unquoted expression to the list of columns to GROUP BY
      * @param string $expr
      */
-    public function group_by_expr($expr) {
+    public function group_by_expr($expr)
+    {
         $this->_group_by[] = $expr;
+
         return $this;
     }
 
@@ -1569,7 +1714,8 @@ class ORM implements ArrayAccess {
      * added, and these will be ANDed together when the final query
      * is built.
      */
-    public function having($column_name, $value) {
+    public function having($column_name, $value)
+    {
         return $this->having_equal($column_name, $value);
     }
 
@@ -1577,7 +1723,8 @@ class ORM implements ArrayAccess {
      * More explicitly named version of for the having() method.
      * Can be used if preferred.
      */
-    public function having_equal($column_name, $value) {
+    public function having_equal($column_name, $value)
+    {
         return $this->_add_simple_having($column_name, '=', $value);
     }
 
@@ -1586,14 +1733,16 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string $value
      */
-    public function having_not_equal($column_name, $value) {
+    public function having_not_equal($column_name, $value)
+    {
         return $this->_add_simple_having($column_name, '!=', $value);
     }
 
     /**
      * Special method to query the table by its primary key
      */
-    public function having_id_is($id) {
+    public function having_id_is($id)
+    {
         return $this->having($this->_get_id_column_name(), $id);
     }
 
@@ -1602,7 +1751,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string $value
      */
-    public function having_like($column_name, $value) {
+    public function having_like($column_name, $value)
+    {
         return $this->_add_simple_having($column_name, 'LIKE', $value);
     }
 
@@ -1611,7 +1761,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string $value
      */
-    public function having_not_like($column_name, $value) {
+    public function having_not_like($column_name, $value)
+    {
         return $this->_add_simple_having($column_name, 'NOT LIKE', $value);
     }
 
@@ -1620,7 +1771,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function having_gt($column_name, $value) {
+    public function having_gt($column_name, $value)
+    {
         return $this->_add_simple_having($column_name, '>', $value);
     }
 
@@ -1629,7 +1781,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function having_lt($column_name, $value) {
+    public function having_lt($column_name, $value)
+    {
         return $this->_add_simple_having($column_name, '<', $value);
     }
 
@@ -1638,7 +1791,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function having_gte($column_name, $value) {
+    public function having_gte($column_name, $value)
+    {
         return $this->_add_simple_having($column_name, '>=', $value);
     }
 
@@ -1647,7 +1801,8 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param integer $value
      */
-    public function having_lte($column_name, $value) {
+    public function having_lte($column_name, $value)
+    {
         return $this->_add_simple_having($column_name, '<=', $value);
     }
 
@@ -1656,9 +1811,11 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string[] $values
      */
-    public function having_in($column_name, $values) {
-        $column_name = $this->_quote_identifier($column_name);
+    public function having_in($column_name, $values)
+    {
+        $column_name  = $this->_quote_identifier($column_name);
         $placeholders = $this->_create_placeholders($values);
+
         return $this->_add_having("{$column_name} IN ({$placeholders})", $values);
     }
 
@@ -1667,9 +1824,11 @@ class ORM implements ArrayAccess {
      * @param string $column_name
      * @param string[] $values
      */
-    public function having_not_in($column_name, $values) {
-        $column_name = $this->_quote_identifier($column_name);
+    public function having_not_in($column_name, $values)
+    {
+        $column_name  = $this->_quote_identifier($column_name);
         $placeholders = $this->_create_placeholders($values);
+
         return $this->_add_having("{$column_name} NOT IN ({$placeholders})", $values);
     }
 
@@ -1677,8 +1836,10 @@ class ORM implements ArrayAccess {
      * Add a HAVING column IS NULL clause to your query
      * @param string $column_name
      */
-    public function having_null($column_name) {
+    public function having_null($column_name)
+    {
         $column_name = $this->_quote_identifier($column_name);
+
         return $this->_add_having("{$column_name} IS NULL");
     }
 
@@ -1686,8 +1847,10 @@ class ORM implements ArrayAccess {
      * Add a HAVING column IS NOT NULL clause to your query
      * @param string $column_name
      */
-    public function having_not_null($column_name) {
+    public function having_not_null($column_name)
+    {
         $column_name = $this->_quote_identifier($column_name);
+
         return $this->_add_having("{$column_name} IS NOT NULL");
     }
 
@@ -1697,7 +1860,8 @@ class ORM implements ArrayAccess {
      * to the parameters supplied in the second argument.
      * @param string $clause
      */
-    public function having_raw($clause, $parameters=array()) {
+    public function having_raw($clause, $parameters = [])
+    {
         return $this->_add_having($clause, $parameters);
     }
 
@@ -1705,17 +1869,19 @@ class ORM implements ArrayAccess {
      * Build a SELECT statement based on the clauses that have
      * been passed to this instance by chaining method calls.
      */
-    protected function _build_select() {
+    protected function _build_select()
+    {
         // If the query is raw, just set the $this->_values to be
         // the raw query parameters and return the raw query
         if ($this->_is_raw_query) {
             $this->_values = $this->_raw_parameters;
+
             return $this->_raw_query;
         }
 
         // Build and return the full SELECT statement by concatenating
         // the results of calling each separate builder method.
-        return $this->_join_if_not_empty(" ", array(
+        return $this->_join_if_not_empty(' ', [
             $this->_build_select_start(),
             $this->_build_join(),
             $this->_build_where(),
@@ -1724,26 +1890,30 @@ class ORM implements ArrayAccess {
             $this->_build_order_by(),
             $this->_build_limit(),
             $this->_build_offset(),
-        ));
+        ]);
     }
 
     /**
      * Used to perform unit tests
      * Note must call _build_select_start() for this to be populated
      */
-    public function testValues() {
+    public function testValues()
+    {
         return $this->_values;
     }
 
     /**
      * Build the start of the SELECT statement
      */
-    protected function _build_select_start() {
-        $fragment = 'SELECT ';
+    protected function _build_select_start()
+    {
+        $fragment       = 'SELECT ';
         $result_columns = join(', ', $this->_result_columns);
 
-        if (!is_null($this->_limit) &&
-            self::$_config[$this->_connection_name]['limit_clause_style'] === ORM::LIMIT_STYLE_TOP_N) {
+        if (
+            !is_null($this->_limit) &&
+            self::$_config[$this->_connection_name]['limit_clause_style'] === ORM::LIMIT_STYLE_TOP_N
+        ) {
             $fragment .= "TOP {$this->_limit} ";
         }
 
@@ -1754,44 +1924,50 @@ class ORM implements ArrayAccess {
         $fragment .= "{$result_columns} FROM " . $this->_quote_identifier($this->_table_name);
 
         if (!is_null($this->_table_alias)) {
-            $fragment .= " " . $this->_quote_identifier($this->_table_alias);
+            $fragment .= ' ' . $this->_quote_identifier($this->_table_alias);
         }
+
         return $fragment;
     }
 
     /**
      * Build the JOIN sources
      */
-    protected function _build_join() {
+    protected function _build_join()
+    {
         if (count($this->_join_sources) === 0) {
             return '';
         }
 
-        return join(" ", $this->_join_sources);
+        return join(' ', $this->_join_sources);
     }
 
     /**
      * Build the WHERE clause(s)
      */
-    protected function _build_where() {
+    protected function _build_where()
+    {
         return $this->_build_conditions('where');
     }
 
     /**
      * Build the HAVING clause(s)
      */
-    protected function _build_having() {
+    protected function _build_having()
+    {
         return $this->_build_conditions('having');
     }
 
     /**
      * Build GROUP BY
      */
-    protected function _build_group_by() {
+    protected function _build_group_by()
+    {
         if (count($this->_group_by) === 0) {
             return '';
         }
-        return "GROUP BY " . join(", ", $this->_group_by);
+
+        return 'GROUP BY ' . join(', ', $this->_group_by);
     }
 
     /**
@@ -1799,39 +1975,45 @@ class ORM implements ArrayAccess {
      * @param string $type
      * @return string
      */
-    protected function _build_conditions($type) {
+    protected function _build_conditions($type)
+    {
         $conditions_class_property_name = "_{$type}_conditions";
         // If there are no clauses, return empty string
         if (count($this->$conditions_class_property_name) === 0) {
             return '';
         }
 
-        $conditions = array();
+        $conditions = [];
         foreach ($this->$conditions_class_property_name as $condition) {
-            $conditions[] = $condition[self::CONDITION_FRAGMENT];
+            $conditions[]  = $condition[self::CONDITION_FRAGMENT];
             $this->_values = array_merge($this->_values, $condition[self::CONDITION_VALUES]);
         }
 
-        return strtoupper($type) . " " . join(" AND ", $conditions);
+        return strtoupper($type) . ' ' . join(' AND ', $conditions);
     }
 
     /**
      * Build ORDER BY
      */
-    protected function _build_order_by() {
+    protected function _build_order_by()
+    {
         if (count($this->_order_by) === 0) {
             return '';
         }
-        return "ORDER BY " . join(", ", $this->_order_by);
+
+        return 'ORDER BY ' . join(', ', $this->_order_by);
     }
 
     /**
      * Build LIMIT
      */
-    protected function _build_limit() {
+    protected function _build_limit()
+    {
         $fragment = '';
-        if (!is_null($this->_limit) &&
-            self::$_config[$this->_connection_name]['limit_clause_style'] == ORM::LIMIT_STYLE_LIMIT) {
+        if (
+            !is_null($this->_limit) &&
+            self::$_config[$this->_connection_name]['limit_clause_style'] == ORM::LIMIT_STYLE_LIMIT
+        ) {
             if (self::$_db[$this->_connection_name]->getAttribute(PDO::ATTR_DRIVER_NAME) == 'firebird') {
                 $fragment = 'ROWS';
             } else {
@@ -1839,20 +2021,24 @@ class ORM implements ArrayAccess {
             }
             $fragment .= " {$this->_limit}";
         }
+
         return $fragment;
     }
 
     /**
      * Build OFFSET
      */
-    protected function _build_offset() {
+    protected function _build_offset()
+    {
         if (!is_null($this->_offset)) {
             $clause = 'OFFSET';
             if (self::$_db[$this->_connection_name]->getAttribute(PDO::ATTR_DRIVER_NAME) == 'firebird') {
                 $clause = 'TO';
             }
+
             return "$clause " . $this->_offset;
         }
+
         return '';
     }
 
@@ -1862,8 +2048,9 @@ class ORM implements ArrayAccess {
      * @param string $glue
      * @return string
      */
-    protected function _join_if_not_empty($glue, $pieces) {
-        $filtered_pieces = array();
+    protected function _join_if_not_empty($glue, $pieces)
+    {
+        $filtered_pieces = [];
         foreach ($pieces as $piece) {
             if (is_string($piece)) {
                 $piece = trim($piece);
@@ -1872,6 +2059,7 @@ class ORM implements ArrayAccess {
                 $filtered_pieces[] = $piece;
             }
         }
+
         return join($glue, $filtered_pieces);
     }
 
@@ -1880,9 +2068,11 @@ class ORM implements ArrayAccess {
      * (table names, column names etc). This method can
      * also deal with dot-separated identifiers eg table.column
      */
-    protected function _quote_identifier($identifier) {
+    protected function _quote_identifier($identifier)
+    {
         $parts = explode('.', $identifier);
-        $parts = array_map(array($this, '_quote_identifier_part'), $parts);
+        $parts = array_map([$this, '_quote_identifier_part'], $parts);
+
         return join('.', $parts);
     }
 
@@ -1891,26 +2081,31 @@ class ORM implements ArrayAccess {
      * part of an identifier, using the identifier quote
      * character specified in the config (or autodetected).
      */
-    protected function _quote_identifier_part($part) {
+    protected function _quote_identifier_part($part)
+    {
         if ($part === '*') {
             return $part;
         }
 
         $quote_character = self::$_config[$this->_connection_name]['identifier_quote_character'];
+
         // double up any identifier quotes to escape them
         return $quote_character .
-               str_replace($quote_character,
-                           $quote_character . $quote_character,
-                           $part
-               ) . $quote_character;
+            str_replace(
+                $quote_character,
+                $quote_character . $quote_character,
+                $part
+            ) . $quote_character;
     }
 
     /**
      * Create a cache key for the given query and parameters.
      */
-    protected static function _create_cache_key($query, $parameters) {
+    protected static function _create_cache_key($query, $parameters)
+    {
         $parameter_string = join(',', $parameters);
-        $key = $query . ':' . $parameter_string;
+        $key              = $query . ':' . $parameter_string;
+
         return sha1($key);
     }
 
@@ -1919,27 +2114,31 @@ class ORM implements ArrayAccess {
      * is cached for the key, return the value. Otherwise, return false.
      * @param string $cache_key
      */
-    protected static function _check_query_cache($cache_key, $connection_name = self::DEFAULT_CONNECTION) {
+    protected static function _check_query_cache($cache_key, $connection_name = self::DEFAULT_CONNECTION)
+    {
         if (isset(self::$_query_cache[$connection_name][$cache_key])) {
             return self::$_query_cache[$connection_name][$cache_key];
         }
+
         return false;
     }
 
     /**
      * Clear the query cache
      */
-    public static function clear_cache() {
-        self::$_query_cache = array();
+    public static function clear_cache()
+    {
+        self::$_query_cache = [];
     }
 
     /**
      * Add the given value to the query cache.
      * @param string $cache_key
      */
-    protected static function _cache_query_result($cache_key, $value, $connection_name = self::DEFAULT_CONNECTION) {
+    protected static function _cache_query_result($cache_key, $value, $connection_name = self::DEFAULT_CONNECTION)
+    {
         if (!isset(self::$_query_cache[$connection_name])) {
-            self::$_query_cache[$connection_name] = array();
+            self::$_query_cache[$connection_name] = [];
         }
         self::$_query_cache[$connection_name][$cache_key] = $value;
     }
@@ -1948,17 +2147,19 @@ class ORM implements ArrayAccess {
      * Execute the SELECT query that has been built up by chaining methods
      * on this class. Return an array of rows as associative arrays.
      */
-    protected function _run() {
-        $query = $this->_build_select();
+    protected function _run()
+    {
+        $query           = $this->_build_select();
         $caching_enabled = self::$_config[$this->_connection_name]['caching'];
 
         $cache_key = '';
         if ($caching_enabled) {
-            $cache_key = self::_create_cache_key($query, $this->_values);
+            $cache_key     = self::_create_cache_key($query, $this->_values);
             $cached_result = self::_check_query_cache($cache_key, $this->_connection_name);
 
             if ($cached_result !== false) {
                 $this->reset();
+
                 return $cached_result;
             }
         }
@@ -1966,7 +2167,7 @@ class ORM implements ArrayAccess {
         self::_execute($query, $this->_values, $this->_connection_name);
         $statement = self::get_last_statement();
 
-        $rows = array();
+        $rows = [];
         while ($row = $statement->fetch(PDO::FETCH_ASSOC)) {
             $rows[] = $row;
         }
@@ -1981,16 +2182,15 @@ class ORM implements ArrayAccess {
         return $rows;
     }
 
-
     /**
      * reset Idiorm after executing the query
-    */
-    protected function reset(){
-        $this->_values = array();
-        $this->_result_columns = array('*');
+     */
+    protected function reset()
+    {
+        $this->_values                       = [];
+        $this->_result_columns               = ['*'];
         $this->_using_default_result_columns = true;
     }
-
 
     /**
      * Return the raw data wrapped by this ORM
@@ -1998,11 +2198,13 @@ class ORM implements ArrayAccess {
      * names may optionally be supplied as arguments,
      * if so, only those keys will be returned.
      */
-    public function as_array() {
+    public function as_array()
+    {
         if (func_num_args() === 0) {
             return $this->_data;
         }
         $args = func_get_args();
+
         return array_intersect_key($this->_data, array_flip($args));
     }
 
@@ -2010,28 +2212,32 @@ class ORM implements ArrayAccess {
      * Return the value of a property of this object (database row)
      * or null if not present.
      */
-    public function get($key) {
-        return isset($this->_data[$key]) ? $this->_data[$key] : null;
+    public function get($key)
+    {
+        return $this->_data[$key] ?? null;
     }
 
     /**
      * Return the name of the column in the database table which contains
      * the primary key ID of the row.
      */
-    protected function _get_id_column_name() {
+    protected function _get_id_column_name()
+    {
         if (!is_null($this->_instance_id_column)) {
             return $this->_instance_id_column;
         }
         if (isset(self::$_config[$this->_connection_name]['id_column_overrides'][$this->_table_name])) {
             return self::$_config[$this->_connection_name]['id_column_overrides'][$this->_table_name];
         }
+
         return self::$_config[$this->_connection_name]['id_column'];
     }
 
     /**
      * Get the primary key ID of this object.
      */
-    public function id() {
+    public function id()
+    {
         return $this->get($this->_get_id_column_name());
     }
 
@@ -2042,7 +2248,8 @@ class ORM implements ArrayAccess {
      * Flags the properties as 'dirty' so they will be saved to the
      * database when save() is called.
      */
-    public function set($key, $value = null) {
+    public function set($key, $value = null)
+    {
         return $this->_set_orm_property($key, $value);
     }
 
@@ -2055,7 +2262,8 @@ class ORM implements ArrayAccess {
      * @param string|array $key
      * @param string|null $value
      */
-    public function set_expr($key, $value = null) {
+    public function set_expr($key, $value = null)
+    {
         return $this->_set_orm_property($key, $value, true);
     }
 
@@ -2064,9 +2272,10 @@ class ORM implements ArrayAccess {
      * @param string|array $key
      * @param string|null $value
      */
-    protected function _set_orm_property($key, $value = null, $expr = false) {
+    protected function _set_orm_property($key, $value = null, $expr = false)
+    {
         if (!is_array($key)) {
-            $key = array($key => $value);
+            $key = [$key => $value];
         }
         foreach ($key as $field => $value) {
             if ($field == '_isFirstResult') {
@@ -2079,12 +2288,16 @@ class ORM implements ArrayAccess {
                 // Save the data the first time only
                 $this->_clean_data[$field] = $this->_data[$field];
             }
-            $oldval = array_key_exists($field, $this->_data) ? $this->_data[$field] : null;
+            $oldval              = array_key_exists($field, $this->_data) ? $this->_data[$field] : null;
             $this->_data[$field] = $value;
-            $set_as_dirty = $this->is_new() || $expr;
+            $set_as_dirty        = $this->is_new() || $expr;
             if (is_float($value)) {
                 $set_as_dirty = abs($oldval - $value) > 0.00000000000001;
-            } else if ($oldval != $value) {
+            } elseif (is_string($oldval)) {
+                if ($oldval !== $value) {
+                    $set_as_dirty = true;
+                }
+            } elseif ($oldval != $value) {
                 $set_as_dirty = true;
             }
             if ($set_as_dirty) {
@@ -2092,10 +2305,11 @@ class ORM implements ArrayAccess {
             }
             if (false === $expr and isset($this->_expr_fields[$field])) {
                 unset($this->_expr_fields[$field]);
-            } else if (true === $expr) {
+            } elseif (true === $expr) {
                 $this->_expr_fields[$field] = true;
             }
         }
+
         return $this;
     }
 
@@ -2105,7 +2319,8 @@ class ORM implements ArrayAccess {
      * @param string $key
      * @return bool
      */
-    public function is_dirty($key) {
+    public function is_dirty($key)
+    {
         return array_key_exists($key, $this->_dirty_fields);
     }
 
@@ -2114,23 +2329,26 @@ class ORM implements ArrayAccess {
      * object was saved.
      * @return bool
      */
-    public function is_any_dirty() {
+    public function is_any_dirty()
+    {
         return count($this->_dirty_fields) > 0;
     }
 
     /**
      * List the dirty fields that need updating on next save
-	 * @return array
+     * @return array
      */
-    public function list_dirty_fields() {
+    public function list_dirty_fields()
+    {
         return $this->_dirty_fields;
     }
 
     /**
      * Get the clean data for this record before it was made dirty
-	 * @return array
+     * @return array
      */
-    public function clean_values() {
+    public function clean_values()
+    {
         return array_merge($this->_data, $this->_clean_data);
     }
 
@@ -2139,21 +2357,24 @@ class ORM implements ArrayAccess {
      * ie before it became dirty
      * @return mixed
      */
-    public function clean_value($key) {
+    public function clean_value($key)
+    {
         if (array_key_exists($key, $this->_clean_data)) {
             return $this->_clean_data[$key];
         }
         if (array_key_exists($key, $this->_data)) {
             return $this->_data[$key];
         }
-        return NULL;
+
+        return null;
     }
 
     /**
      * Check whether the model was the result of a call to create() or not
      * @return bool
      */
-    public function is_new() {
+    public function is_new()
+    {
         return $this->_is_new;
     }
 
@@ -2163,8 +2384,9 @@ class ORM implements ArrayAccess {
      * Added: on duplicate key update, only for mysql
      * If you want to insert a record, or update it if any of the unique keys already exists on db
      */
-    public function save($ignore = false) {
-        $query = array();
+    public function save($ignore = false)
+    {
+        $query = [];
 
         // Fix if id field is blank but not null
         if (!($this->id() && array_key_exists($this->_get_id_column_name(), $this->_dirty_fields))) {
@@ -2174,19 +2396,16 @@ class ORM implements ArrayAccess {
         // remove any expression fields as they are already baked into the query
         $values = array_values(array_diff_key($this->_dirty_fields, $this->_expr_fields));
 
-        if($ignore)
-        {
-            $query   = $this->_build_insert_update();
-            $values  = array_merge($values, $values);
-        }
-        else
-        {
+        if ($ignore) {
+            $query  = $this->_build_insert_update();
+            $values = array_merge($values, $values);
+        } else {
             if (!$this->_is_new) { // UPDATE
                 // If there are no dirty values, do nothing
                 if (empty($values) && empty($this->_expr_fields)) {
                     return true;
                 }
-                $query = $this->_build_update();
+                $query    = $this->_build_update();
                 $values[] = $this->id();
             } else { // INSERT
                 $query = $this->_build_insert();
@@ -2199,7 +2418,7 @@ class ORM implements ArrayAccess {
         if ($this->_is_new) {
             $this->_is_new = false;
             if (!($this->id())) {
-                if(self::$_db[$this->_connection_name]->getAttribute(PDO::ATTR_DRIVER_NAME) == 'pgsql') {
+                if (self::$_db[$this->_connection_name]->getAttribute(PDO::ATTR_DRIVER_NAME) == 'pgsql') {
                     $this->_data[$this->_get_id_column_name()] = self::get_last_statement()->fetchColumn();
                 } else {
                     $this->_data[$this->_get_id_column_name()] = self::$_db[$this->_connection_name]->lastInsertId();
@@ -2207,86 +2426,92 @@ class ORM implements ArrayAccess {
             }
         }
         $this->clear_cache();
-        $this->_dirty_fields = $this->_expr_fields = array();
-        $this->_clean_data = array();
+        $this->_dirty_fields = $this->_expr_fields = [];
+        $this->_clean_data   = [];
+
         return $success;
     }
 
     /**
      * Build an UPDATE query
      */
-    protected function _build_update() {
-        $query = array();
+    protected function _build_update()
+    {
+        $query   = [];
         $query[] = "UPDATE {$this->_quote_identifier($this->_table_name)} SET";
 
-        $field_list = array();
+        $field_list = [];
         foreach ($this->_dirty_fields as $key => $value) {
-            if(!array_key_exists($key, $this->_expr_fields)) {
+            if (!array_key_exists($key, $this->_expr_fields)) {
                 $value = '?';
             }
             $field_list[] = "{$this->_quote_identifier($key)} = $value";
         }
-        $query[] = join(", ", $field_list);
-        $query[] = "WHERE";
+        $query[] = join(', ', $field_list);
+        $query[] = 'WHERE';
         $query[] = $this->_quote_identifier($this->_get_id_column_name());
-        $query[] = "= ?";
-        return join(" ", $query);
+        $query[] = '= ?';
+
+        return join(' ', $query);
     }
 
     /**
      * Build an INSERT query
      */
-    protected function _build_insert() {
-        $query = array();
-        $query[] = "INSERT INTO";
-        $query[] = $this->_quote_identifier($this->_table_name);
-        $field_list = array_map(array($this, '_quote_identifier'), array_keys($this->_dirty_fields));
-        $query[] = "(" . join(", ", $field_list) . ")";
-        $query[] = "VALUES";
+    protected function _build_insert()
+    {
+        $query      = [];
+        $query[]    = 'INSERT INTO';
+        $query[]    = $this->_quote_identifier($this->_table_name);
+        $field_list = array_map([$this, '_quote_identifier'], array_keys($this->_dirty_fields));
+        $query[]    = '(' . join(', ', $field_list) . ')';
+        $query[]    = 'VALUES';
 
         $placeholders = $this->_create_placeholders($this->_dirty_fields);
-        $query[] = "({$placeholders})";
+        $query[]      = "({$placeholders})";
 
         if (self::$_db[$this->_connection_name]->getAttribute(PDO::ATTR_DRIVER_NAME) == 'pgsql') {
             $query[] = 'RETURNING ' . $this->_quote_identifier($this->_get_id_column_name());
         }
 
-        return join(" ", $query);
+        return join(' ', $query);
     }
 
     /**
      * Added: Build an INSERT ON DUPLICATE KEY UPDATE query
      * Attention: This method only works on Mysql Databases
      */
-    protected function _build_insert_update() {
-        $query = array();
-        $query[] = "INSERT INTO";
-        $query[] = $this->_quote_identifier($this->_table_name);
-        $field_list = array_map(array($this, '_quote_identifier'), array_keys($this->_dirty_fields));
-        $query[] = "(" . implode(", ", $field_list) . ")";
-        $query[] = "VALUES";
+    protected function _build_insert_update()
+    {
+        $query        = [];
+        $query[]      = 'INSERT INTO';
+        $query[]      = $this->_quote_identifier($this->_table_name);
+        $field_list   = array_map([$this, '_quote_identifier'], array_keys($this->_dirty_fields));
+        $query[]      = '(' . implode(', ', $field_list) . ')';
+        $query[]      = 'VALUES';
         $placeholders = $this->_create_placeholders($this->_dirty_fields);
-        $query[] = "({$placeholders})";
+        $query[]      = "({$placeholders})";
 
-        $query[] = " ON DUPLICATE KEY UPDATE ";
-        $query[] =  implode(" = ?, ", $field_list)." = ? ";
-        return implode(" ", $query);
+        $query[] = ' ON DUPLICATE KEY UPDATE ';
+        $query[] = implode(' = ?, ', $field_list) . ' = ? ';
+
+        return implode(' ', $query);
     }
-
 
     /**
      * Delete this record from the database
      */
-    public function delete() {
-        $query = join(" ", array(
-            "DELETE FROM",
+    public function delete()
+    {
+        $query = join(' ', [
+            'DELETE FROM',
             $this->_quote_identifier($this->_table_name),
-            "WHERE",
+            'WHERE',
             $this->_quote_identifier($this->_get_id_column_name()),
-            "= ?",
-        ));
+            '= ?',
+        ]);
 
-        return self::_execute($query, array($this->id()), $this->_connection_name);
+        return self::_execute($query, [$this->id()], $this->_connection_name);
     }
 
     /**
@@ -2294,54 +2519,53 @@ class ORM implements ArrayAccess {
      * Added: could delete many of a join query, if you define $join to true
      * and the table where you want to delete the records
      */
-    public function delete_many($join = false, $table = false) {
-        if($join){
+    public function delete_many($join = false, $table = false)
+    {
+        if ($join) {
             // Build and return the full DELETE statement by concatenating
             // the results of calling each separate builder method.
-            $query = $this->_join_if_not_empty(" ", array(
+            $query = $this->_join_if_not_empty(' ', [
                 "DELETE $table FROM",
                 $this->_quote_identifier($this->_table_name),
                 $this->_build_join(),
-                $this->_build_where()
-            ));
-        }
-        else{
+                $this->_build_where(),
+            ]);
+        } else {
             // Build and return the full DELETE statement by concatenating
             // the results of calling each separate builder method.
-            $query = $this->_join_if_not_empty(" ", array(
-                "DELETE FROM",
+            $query = $this->_join_if_not_empty(' ', [
+                'DELETE FROM',
                 $this->_quote_identifier($this->_table_name),
                 $this->_build_where(),
-            ));
+            ]);
         }
 
         return self::_execute($query, $this->_values, $this->_connection_name);
     }
 
-
     // --------------------- //
     // ---  ArrayAccess  --- //
     // --------------------- //
-    #[\ReturnTypeWillChange]
-    public function offsetExists($key) {
+    public function offsetExists($key): bool
+    {
         return isset($this->_data[$key]);
     }
 
-    #[\ReturnTypeWillChange]
-    public function offsetGet($key) {
+    public function offsetGet($key): mixed
+    {
         return $this->get($key);
     }
 
-    #[\ReturnTypeWillChange]
-    public function offsetSet($key, $value) {
-        if(is_null($key)) {
+    public function offsetSet($key, $value): void
+    {
+        if (is_null($key)) {
             throw new \InvalidArgumentException('You must specify a key/array index.');
         }
         $this->set($key, $value);
     }
 
-    #[\ReturnTypeWillChange]
-    public function offsetUnset($key) {
+    public function offsetUnset($key): void
+    {
         unset($this->_data[$key]);
         unset($this->_dirty_fields[$key]);
     }
@@ -2349,20 +2573,23 @@ class ORM implements ArrayAccess {
     // --------------------- //
     // --- MAGIC METHODS --- //
     // --------------------- //
-    public function __get($key) {
+    public function __get($key)
+    {
         return $this->offsetGet($key);
     }
 
-    public function __set($key, $value) {
+    public function __set($key, $value)
+    {
         $this->offsetSet($key, $value);
     }
 
-    public function __unset($key) {
+    public function __unset($key)
+    {
         $this->offsetUnset($key);
     }
 
-
-    public function __isset($key) {
+    public function __isset($key)
+    {
         return $this->offsetExists($key);
     }
 
@@ -2381,13 +2608,13 @@ class ORM implements ArrayAccess {
     public function __call($name, $arguments)
     {
         $method = strtolower(preg_replace('/([a-z])([A-Z])/', '$1_$2', $name));
-        //return call_user_func_array(array($this, $method), $arguments);
+        // return call_user_func_array(array($this, $method), $arguments);
 
         if (method_exists($this, $method)) {
-            return call_user_func_array(array($this, $method), $arguments);
-        } else {
-            return false;
+            return call_user_func_array([$this, $method], $arguments);
         }
+
+        return false;
     }
 
     /**
@@ -2406,6 +2633,6 @@ class ORM implements ArrayAccess {
     {
         $method = strtolower(preg_replace('/([a-z])([A-Z])/', '$1_$2', $name));
 
-        return call_user_func_array(array('self', $method), $arguments);
+        return call_user_func_array(['self', $method], $arguments);
     }
 }
