@@ -12,19 +12,19 @@ use Granada\Model;
  * Additions:
  *  - Test will also check for double records on a has_many relation
  */
-class EagerTest extends PHPUnit_Framework_TestCase {
-
+class EagerTest extends \PHPUnit\Framework\TestCase
+{
     /**
      * @before
      */
-    protected function beforeTest() {
-
+    protected function beforeTest()
+    {
         // The tests for eager loading requires a real database.
         // Set up SQLite in memory
         ORM::set_db(new PDO('sqlite::memory:'));
 
         // Create schemas and populate with data
-        ORM::get_db()->exec(file_get_contents(__DIR__ . DIRECTORY_SEPARATOR . '..'.DIRECTORY_SEPARATOR.'models.sql'));
+        ORM::get_db()->exec(file_get_contents(__DIR__ . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . 'models.sql'));
 
         // Enable logging
         ORM::configure('logging', true);
@@ -33,17 +33,17 @@ class EagerTest extends PHPUnit_Framework_TestCase {
     /**
      * @after
      */
-    protected function afterTest() {
+    protected function afterTest()
+    {
         ORM::configure('logging', false);
         ORM::set_db(null);
     }
 
-
-    public function testFindOneWith1BelongsTo() {
-
+    public function testFindOneWith1BelongsTo()
+    {
         $car = Model::factory('Car')->with('manufactor')->find_one(1);
 
-        $expectedSql   = array();
+        $expectedSql   = [];
         $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `id` = '1' LIMIT 1";
         $expectedSql[] = "SELECT * FROM `manufactor` WHERE `enabled` = '1' AND `id` IN ('1')";
 
@@ -53,13 +53,13 @@ class EagerTest extends PHPUnit_Framework_TestCase {
         $actualSql = array_slice($fullQueryLog, count($fullQueryLog) - 2);
 
         $this->assertEquals($expectedSql, $actualSql);
-
     }
 
-    public function testFindOneWith2BelongsTo() {
-        $car = Model::factory('Car')->with('owner','manufactor')->find_one(1);
+    public function testFindOneWith2BelongsTo()
+    {
+        $car = Model::factory('Car')->with('owner', 'manufactor')->find_one(1);
 
-        $expectedSql = array();
+        $expectedSql   = [];
         $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `id` = '1' LIMIT 1";
         $expectedSql[] = "SELECT * FROM `owner` WHERE `id` IN ('1')";
         $expectedSql[] = "SELECT * FROM `manufactor` WHERE `enabled` = '1' AND `id` IN ('1')";
@@ -72,10 +72,11 @@ class EagerTest extends PHPUnit_Framework_TestCase {
         $this->assertEquals($expectedSql, $actualSql);
     }
 
-    public function testFindOneWith1HasOne() {
+    public function testFindOneWith1HasOne()
+    {
         $owner = Model::factory('Owner')->with('car')->find_one(1);
 
-        $expectedSql   = array();
+        $expectedSql   = [];
         $expectedSql[] = "SELECT * FROM `owner` WHERE `id` = '1' LIMIT 1";
         $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `owner_id` IN ('1')";
 
@@ -87,13 +88,14 @@ class EagerTest extends PHPUnit_Framework_TestCase {
         $this->assertEquals($expectedSql, $actualSql);
     }
 
-    public function testFindOneWith1HasMany() {
+    public function testFindOneWith1HasMany()
+    {
         $manufactor = Model::factory('Manufactor')->with('cars')->find_one(1);
 
-        $expectedSql   = array();
+        $expectedSql   = [];
         $expectedSql[] = "SELECT * FROM `manufactor` WHERE `id` = '1' LIMIT 1";
         $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `enabled` = '1' AND `manufactor_id` IN ('1')";
-        $fullQueryLog = ORM::get_query_log();
+        $fullQueryLog  = ORM::get_query_log();
 
         // Return last two queries
         $actualSql = array_slice($fullQueryLog, count($fullQueryLog) - 2);
@@ -101,37 +103,37 @@ class EagerTest extends PHPUnit_Framework_TestCase {
         $this->assertEquals($expectedSql, $actualSql);
     }
 
-     public function testFindOneWithHasManyThrough() {
-        $car = Model::factory('Car')->with('parts')->find_one(1);
-        $actualParts = array();
-        foreach($car->parts as $p) {
+    public function testFindOneWithHasManyThrough()
+    {
+        $car         = Model::factory('Car')->with('parts')->find_one(1);
+        $actualParts = [];
+        foreach ($car->parts as $p) {
             $actualParts[] = $p->as_array();
         }
 
-        $expectedSql    = array();
-        $expectedSql[]  = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `id` = '1' LIMIT 1";
-        $expectedSql[]  = "SELECT `part`.*, `car_part`.`car_id` FROM `part` JOIN `car_part` ON `part`.`id` = `car_part`.`part_id` WHERE `car_part`.`car_id` IN ('1')";
+        $expectedSql   = [];
+        $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `id` = '1' LIMIT 1";
+        $expectedSql[] = "SELECT `part`.*, `car_part`.`car_id` FROM `part` JOIN `car_part` ON `part`.`id` = `car_part`.`part_id` WHERE `car_part`.`car_id` IN ('1')";
 
-        $expectedParts = array();
-        $expectedParts[] = array('id' => 1, 'name' => 'Part1');
-        $expectedParts[] = array('id' => 2, 'name' => 'Part2');
-        $expectedParts[] = array('id' => 1, 'name' => 'Part1');
+        $expectedParts   = [];
+        $expectedParts[] = ['id' => 1, 'name' => 'Part1'];
+        $expectedParts[] = ['id' => 2, 'name' => 'Part2'];
+        $expectedParts[] = ['id' => 1, 'name' => 'Part1'];
 
         $fullQueryLog = ORM::get_query_log();
 
         // Return last three queries
         $actualSql = array_slice($fullQueryLog, count($fullQueryLog) - 2);
 
-
         $this->assertEquals($expectedSql, $actualSql);
         $this->assertEquals($expectedParts, $actualParts);
-
     }
 
-    public function testFindManyWith2BelongsTo() {
-        $cars = Model::factory('Car')->with('owner','manufactor')->find_many();
+    public function testFindManyWith2BelongsTo()
+    {
+        $cars = Model::factory('Car')->with('owner', 'manufactor')->find_many();
 
-        $expectedSql = array();
+        $expectedSql   = [];
         $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0'";
         $expectedSql[] = "SELECT * FROM `owner` WHERE `id` IN ('1', '2', '3', '4')";
         $expectedSql[] = "SELECT * FROM `manufactor` WHERE `enabled` = '1' AND `id` IN ('1', '2')";
@@ -144,11 +146,12 @@ class EagerTest extends PHPUnit_Framework_TestCase {
         $this->assertEquals($expectedSql, $actualSql);
     }
 
-    public function testFindManyWith1HasOne() {
+    public function testFindManyWith1HasOne()
+    {
         $owner = Model::factory('Owner')->with('car')->find_many();
 
-        $expectedSql   = array();
-        $expectedSql[] = "SELECT * FROM `owner`";
+        $expectedSql   = [];
+        $expectedSql[] = 'SELECT * FROM `owner`';
         $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `owner_id` IN ('1', '2', '3', '4')";
 
         $fullQueryLog = ORM::get_query_log();
@@ -159,11 +162,12 @@ class EagerTest extends PHPUnit_Framework_TestCase {
         $this->assertEquals($expectedSql, $actualSql);
     }
 
-    public function testFindManyWith1HasMany() {
+    public function testFindManyWith1HasMany()
+    {
         $manufactor = Model::factory('Manufactor')->with('cars')->find_many();
 
-        $expectedSql   = array();
-        $expectedSql[] = "SELECT * FROM `manufactor`";
+        $expectedSql   = [];
+        $expectedSql[] = 'SELECT * FROM `manufactor`';
         $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `enabled` = '1' AND `manufactor_id` IN ('1', '2')";
 
         $fullQueryLog = ORM::get_query_log();
@@ -174,63 +178,73 @@ class EagerTest extends PHPUnit_Framework_TestCase {
         $this->assertEquals($expectedSql, $actualSql);
     }
 
-    public function testFindManyWithHasManyThrough() {
+    public function testFindManyWithHasManyThrough()
+    {
         $parts = Model::factory('Part')->with('cars')->find_many();
 
-        $actualParts = array();
+        $actualParts = [];
 
-        foreach($parts as $part) {
-            $tmp = $part->as_array();
-            $tmp['cars'] = array();
+        foreach ($parts as $part) {
+            $tmp         = $part->as_array();
+            $tmp['cars'] = [];
 
-            if(count($part->cars) > 0) {
-                foreach($part->cars as $car) {
+            if (count($part->cars) > 0) {
+                foreach ($part->cars as $car) {
                     $tmp['cars'][] = $car->as_array();
-
                 }
             }
 
             $actualParts[] = $tmp;
         }
 
-        $expectedParts = array();
-        $expectedParts[] =  array('id' => '1', 'name' => 'Part1',
-                                'cars' => array(
-                                    array('id' => '1', 'name' => 'Car1', 'manufactor_id' => '1', 'owner_id' => '1', 'is_deleted' => '0', 'enabled' => '1'),
-                                    array('id' => '2', 'name' => 'Car2', 'manufactor_id' => '1', 'owner_id' => '2', 'is_deleted' => '0', 'enabled' => '1'),
-                                    array('id' => '3', 'name' => 'Car3', 'manufactor_id' => '2', 'owner_id' => '3', 'is_deleted' => '0', 'enabled' => '1'),
-                                    array('id' => '4', 'name' => 'Car4', 'manufactor_id' => '2', 'owner_id' => '4', 'is_deleted' => '0', 'enabled' => '1'),
-                                    array('id' => '1', 'name' => 'Car1', 'manufactor_id' => '1', 'owner_id' => '1', 'is_deleted' => '0', 'enabled' => '1'),
-                                )
-                            );
+        $expectedParts   = [];
+        $expectedParts[] = [
+            'id'   => '1',
+            'name' => 'Part1',
+            'cars' => [
+                ['id' => '1', 'name' => 'Car1', 'manufactor_id' => '1', 'owner_id' => '1', 'is_deleted' => '0', 'enabled' => '1'],
+                ['id' => '2', 'name' => 'Car2', 'manufactor_id' => '1', 'owner_id' => '2', 'is_deleted' => '0', 'enabled' => '1'],
+                ['id' => '3', 'name' => 'Car3', 'manufactor_id' => '2', 'owner_id' => '3', 'is_deleted' => '0', 'enabled' => '1'],
+                ['id' => '4', 'name' => 'Car4', 'manufactor_id' => '2', 'owner_id' => '4', 'is_deleted' => '0', 'enabled' => '1'],
+                ['id' => '1', 'name' => 'Car1', 'manufactor_id' => '1', 'owner_id' => '1', 'is_deleted' => '0', 'enabled' => '1'],
+            ],
+        ];
 
-        $expectedParts[] =  array('id' => '2', 'name' => 'Part2',
-                                'cars' => array(
-                                    array('id' => '1', 'name' => 'Car1', 'manufactor_id' => '1', 'owner_id' => '1', 'is_deleted' => '0', 'enabled' => '1'),
-                                )
-                            );
+        $expectedParts[] = [
+            'id'   => '2',
+            'name' => 'Part2',
+            'cars' => [
+                ['id' => '1', 'name' => 'Car1', 'manufactor_id' => '1', 'owner_id' => '1', 'is_deleted' => '0', 'enabled' => '1'],
+            ],
+        ];
 
-        $expectedParts[] =  array('id' => '3', 'name' => 'Part3',
-                                'cars' => array(
-                                    array('id' => '2', 'name' => 'Car2', 'manufactor_id' => '1', 'owner_id' => '2', 'is_deleted' => '0', 'enabled' => '1'),
-                                )
-                            );
+        $expectedParts[] = [
+            'id'   => '3',
+            'name' => 'Part3',
+            'cars' => [
+                ['id' => '2', 'name' => 'Car2', 'manufactor_id' => '1', 'owner_id' => '2', 'is_deleted' => '0', 'enabled' => '1'],
+            ],
+        ];
 
-        $expectedParts[] =  array('id' =>  '4', 'name' => 'Part4',
-                                'cars' => array(
-                                    array('id' => '3', 'name' => 'Car3', 'manufactor_id' => '2', 'owner_id' => '3', 'is_deleted' => '0', 'enabled' => '1'),
-                                )
-                            );
+        $expectedParts[] = [
+            'id'   => '4',
+            'name' => 'Part4',
+            'cars' => [
+                ['id' => '3', 'name' => 'Car3', 'manufactor_id' => '2', 'owner_id' => '3', 'is_deleted' => '0', 'enabled' => '1'],
+            ],
+        ];
 
-        $expectedParts[] =  array('id' => '5', 'name' => 'Part5',
-                                'cars' => array(
-                                    array('id' => '4', 'name' => 'Car4', 'manufactor_id' => '2', 'owner_id' => '4', 'is_deleted' => '0', 'enabled' => '1'),
-                                )
-                            );
+        $expectedParts[] = [
+            'id'   => '5',
+            'name' => 'Part5',
+            'cars' => [
+                ['id' => '4', 'name' => 'Car4', 'manufactor_id' => '2', 'owner_id' => '4', 'is_deleted' => '0', 'enabled' => '1'],
+            ],
+        ];
 
-        $expectedSql    = array();
-        $expectedSql[]  = "SELECT * FROM `part`";
-        $expectedSql[]  = "SELECT `car`.*, `car_part`.`part_id` FROM `car` JOIN `car_part` ON `car`.`id` = `car_part`.`car_id` WHERE `car`.`is_deleted` = '0' AND `car_part`.`part_id` IN ('1', '2', '3', '4', '5')";
+        $expectedSql   = [];
+        $expectedSql[] = 'SELECT * FROM `part`';
+        $expectedSql[] = "SELECT `car`.*, `car_part`.`part_id` FROM `car` JOIN `car_part` ON `car`.`id` = `car_part`.`car_id` WHERE `car`.`is_deleted` = '0' AND `car_part`.`part_id` IN ('1', '2', '3', '4', '5')";
 
         $fullQueryLog = ORM::get_query_log();
 
@@ -239,43 +253,46 @@ class EagerTest extends PHPUnit_Framework_TestCase {
 
         $this->assertEquals($expectedSql, $actualSql);
         $this->assertEquals($expectedParts, $actualParts);
-
     }
 
-    public function testChainedRelationships() {
-        $owner = Owner::with(array('car'=>array('with'=>'manufactor')))->find_one(1);
+    public function testChainedRelationships()
+    {
+        $owner        = Owner::with(['car' => ['with' => 'manufactor']])->find_one(1);
         $fullQueryLog = ORM::get_query_log();
         // Return last three queries
         $actualSql = array_slice($fullQueryLog, count($fullQueryLog) - 3);
 
-        $expectedSql    = array();
-        $expectedSql[]  = "SELECT * FROM `owner` WHERE `id` = '1' LIMIT 1";
-        $expectedSql[]  = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `owner_id` IN ('1')";
-        $expectedSql[]  = "SELECT * FROM `manufactor` WHERE `enabled` = '1' AND `id` IN ('1')";
+        $expectedSql   = [];
+        $expectedSql[] = "SELECT * FROM `owner` WHERE `id` = '1' LIMIT 1";
+        $expectedSql[] = "SELECT * FROM `car` WHERE `car`.`is_deleted` = '0' AND `owner_id` IN ('1')";
+        $expectedSql[] = "SELECT * FROM `manufactor` WHERE `enabled` = '1' AND `id` IN ('1')";
 
         $this->assertEquals($expectedSql, $actualSql);
     }
 
-    public function testChainedAdterHas_Many_Through() {
-        $car = Car::with(array('parts'=>array('with'=>'cars')))->find_one(1);
+    public function testChainedAdterHas_Many_Through()
+    {
+        $car         = Car::with(['parts' => ['with' => 'cars']])->find_one(1);
         $test_exists = $car->as_array();
         $test_exists = $car->parts->as_array();
-        foreach($car->parts as $part){
-            foreach($part->cars as $car){
+        foreach ($car->parts as $part) {
+            foreach ($part->cars as $car) {
                 $test_exists = $car->as_array();
-            };
+            }
         }
         // NO FATAL ERRORS OR EXCEPTIONS THROW
         $this->assertInstanceOf('Granada\Model', $car);
     }
 
-    public function testLazyLoading() {
+    public function testLazyLoading()
+    {
         $owner = Model::factory('Owner')->find_one(1);
         $this->assertEquals($owner->car->manufactor_id, 1);
     }
 
-    public function testLazyMissingManufacturerNull() {
-        $man1 = Manufactor::find_one(1);
+    public function testLazyMissingManufacturerNull()
+    {
+        $man1          = Manufactor::find_one(1);
         $man1->enabled = false;
         $man1->save();
 
@@ -284,8 +301,9 @@ class EagerTest extends PHPUnit_Framework_TestCase {
         $this->assertNull($car1->manufactor);
     }
 
-    public function testEagerMissingManufacturerNull() {
-        $man1 = Manufactor::find_one(1);
+    public function testEagerMissingManufacturerNull()
+    {
+        $man1          = Manufactor::find_one(1);
         $man1->enabled = false;
         $man1->save();
 
