@@ -1,14 +1,14 @@
 <?php
 
 use Granada\ORM;
-use Granada\ResultSet;
 
-class ConfigTest extends PHPUnit_Framework_TestCase {
-
+class ConfigTest extends \PHPUnit\Framework\TestCase
+{
     /**
      * @before
      */
-    protected function beforeTest() {
+    protected function beforeTest()
+    {
         // Enable logging
         ORM::configure('logging', true);
 
@@ -22,29 +22,34 @@ class ConfigTest extends PHPUnit_Framework_TestCase {
     /**
      * @after
      */
-    protected function afterTest() {
+    protected function afterTest()
+    {
         ORM::reset_config();
         ORM::reset_db();
     }
 
-    protected function setUpIdColumnOverrides() {
-        ORM::configure('id_column_overrides', array(
-            'widget' => 'widget_id',
+    protected function setUpIdColumnOverrides()
+    {
+        ORM::configure('id_column_overrides', [
+            'widget'        => 'widget_id',
             'widget_handle' => 'widget_handle_id',
-        ));
+        ]);
     }
 
-    protected function tearDownIdColumnOverrides() {
-        ORM::configure('id_column_overrides', array());
+    protected function tearDownIdColumnOverrides()
+    {
+        ORM::configure('id_column_overrides', []);
     }
 
-    public function testSettingIdColumn() {
+    public function testSettingIdColumn()
+    {
         ORM::for_table('widget')->find_one(5);
         $expected = "SELECT * FROM `widget` WHERE `primary_key` = '5' LIMIT 1";
         $this->assertEquals($expected, ORM::get_last_query());
     }
 
-    public function testSettingIdColumnOverridesOne() {
+    public function testSettingIdColumnOverridesOne()
+    {
         $this->setUpIdColumnOverrides();
 
         ORM::for_table('widget')->find_one(5);
@@ -54,7 +59,8 @@ class ConfigTest extends PHPUnit_Framework_TestCase {
         $this->tearDownIdColumnOverrides();
     }
 
-    public function testSettingIdColumnOverridesTwo() {
+    public function testSettingIdColumnOverridesTwo()
+    {
         $this->setUpIdColumnOverrides();
 
         ORM::for_table('widget_handle')->find_one(5);
@@ -64,7 +70,8 @@ class ConfigTest extends PHPUnit_Framework_TestCase {
         $this->tearDownIdColumnOverrides();
     }
 
-    public function testSettingIdColumnOverridesThree() {
+    public function testSettingIdColumnOverridesThree()
+    {
         $this->setUpIdColumnOverrides();
 
         ORM::for_table('widget_nozzle')->find_one(5);
@@ -74,7 +81,8 @@ class ConfigTest extends PHPUnit_Framework_TestCase {
         $this->tearDownIdColumnOverrides();
     }
 
-    public function testInstanceIdColumnOne() {
+    public function testInstanceIdColumnOne()
+    {
         $this->setUpIdColumnOverrides();
 
         ORM::for_table('widget')->use_id_column('new_id')->find_one(5);
@@ -84,7 +92,8 @@ class ConfigTest extends PHPUnit_Framework_TestCase {
         $this->tearDownIdColumnOverrides();
     }
 
-    public function testInstanceIdColumnTwo() {
+    public function testInstanceIdColumnTwo()
+    {
         $this->setUpIdColumnOverrides();
 
         ORM::for_table('widget_handle')->use_id_column('new_id')->find_one(5);
@@ -94,7 +103,8 @@ class ConfigTest extends PHPUnit_Framework_TestCase {
         $this->tearDownIdColumnOverrides();
     }
 
-    public function testInstanceIdColumnThree() {
+    public function testInstanceIdColumnThree()
+    {
         $this->setUpIdColumnOverrides();
 
         ORM::for_table('widget_nozzle')->use_id_column('new_id')->find_one(5);
@@ -104,31 +114,32 @@ class ConfigTest extends PHPUnit_Framework_TestCase {
         $this->tearDownIdColumnOverrides();
     }
 
-    public function testGetConfig() {
+    public function testGetConfig()
+    {
         $this->assertTrue(ORM::get_config('logging'));
         ORM::configure('logging', false);
         $this->assertFalse(ORM::get_config('logging'));
         ORM::configure('logging', true);
     }
 
-    public function testGetConfigArray() {
-        $expected = array(
-            'connection_string' => 'sqlite::memory:',
-            'id_column' => 'primary_key',
-            'id_column_overrides' => array(),
-            'error_mode' => PDO::ERRMODE_EXCEPTION,
-            'username' => null,
-            'password' => null,
-            'driver_options' => null,
-            'identifier_quote_character' => '`',
-            'logging' => true,
-            'logger' => null,
-            'caching' => false,
-            'return_result_sets' => true, // true by default in Granada
-            'limit_clause_style' => 'limit',
+    public function testGetConfigArray()
+    {
+        $expected = [
+            'connection_string'           => 'sqlite::memory:',
+            'id_column'                   => 'primary_key',
+            'id_column_overrides'         => [],
+            'error_mode'                  => PDO::ERRMODE_EXCEPTION,
+            'username'                    => null,
+            'password'                    => null,
+            'driver_options'              => null,
+            'identifier_quote_character'  => '`',
+            'logging'                     => true,
+            'logger'                      => null,
+            'caching'                     => false,
+            'return_result_sets'          => true, // true by default in Granada
+            'limit_clause_style'          => 'limit',
             'find_many_primary_id_as_key' => true,
-        );
+        ];
         $this->assertEquals($expected, ORM::get_config());
     }
-
 }
